refactor(#1104): split inbox.rs (4,371 LOC) into inbox/ module directory
Extracts 5 submodules from the 2nd largest file in the codebase:
- message.rs — InboxMessage, ForceMeta, BroadcastContext, NotifySource, MessageStatus
- storage.rs — enqueue, drain, sweep_expired, query functions, path resolution
- notify.rs — PTY injection, compose_aware_inject, dedup gating, format_header
- disk.rs — disk health, readonly mode, half-write recovery
- tests.rs — all 104 tests via #[path] pattern

mod.rs re-exports maintain full API compatibility — no caller changes needed.
.gitignore updated: /inbox/ (root only) to not shadow src/inbox/.

Fixes #1104


Agend-Agent: fixup-dev
Agend-Task: t-20260524004157980227-25
Agend-Branch: refactor/1104-inbox-module-split
Agend-Issued-At: 2026-05-24T00:42:55.156894+00:00
