--- src/inbox.rs
@@ -1,4370 +1,0 @@
-//! Per-agent message inbox — append-only JSONL with disk resilience.
-//!
-//! Messages stored as one JSON object per line in {home}/inbox/{name}.jsonl.
-//!
-//! Resilience layers:
-//! - **Readonly mode**: when available disk space < 5%, enqueue returns an
-//!   error while drain continues to work (let agents consume backlog).
-//! - **Atomic append**: each enqueue writes to a temp file, fsyncs, then
-//!   renames — no half-written lines on crash.
-//! - **Half-write recovery**: on startup, stale `.tmp` files and corrupt
-//!   JSONL lines are moved to `inbox.recovery/` for forensics.
-
-use serde::{Deserialize, Serialize};
-use std::borrow::Cow;
-use std::fmt;
-use std::io::Write;
-use std::path::{Path, PathBuf};
-use std::sync::atomic::{AtomicBool, Ordering};
-
-/// Global readonly flag — set when available disk space drops below threshold.
-static DISK_READONLY: AtomicBool = AtomicBool::new(false);
-
-/// Minimum free-space ratio before entering readonly mode.
-const LOW_DISK_THRESHOLD: f64 = 0.05;
-
-/// Check available disk space at `path`. Returns true if below threshold.
-fn is_disk_low(path: &Path) -> bool {
-    use fs4::available_space;
-    use fs4::total_space;
-    let avail = match available_space(path) {
-        Ok(s) => s,
-        Err(_) => return false, // can't check → assume OK
-    };
-    let total = match total_space(path) {
-        Ok(s) if s > 0 => s,
-        _ => return false,
-    };
-    (avail as f64 / total as f64) < LOW_DISK_THRESHOLD
-}
-
-/// Update the global readonly flag based on disk space at `home`.
-/// Called at daemon startup and before each enqueue.
-pub fn check_disk_space(home: &Path) {
-    let readonly = is_disk_low(home);
-    let was = DISK_READONLY.swap(readonly, Ordering::Relaxed);
-    if readonly && !was {
-        tracing::warn!("inbox entering readonly mode — disk space < 5%");
-    } else if !readonly && was {
-        tracing::info!("inbox leaving readonly mode — disk space recovered");
-    }
-}
-
-/// Returns true when inbox is in readonly mode (disk full).
-pub fn is_readonly() -> bool {
-    DISK_READONLY.load(Ordering::Relaxed)
-}
-
-/// Scan the inbox directory for stale `.tmp` files and corrupt JSONL,
-/// moving them to `inbox.recovery/<timestamp>/` for forensics.
-/// Call once at daemon startup.
-pub fn recover_half_writes(home: &Path) {
-    let inbox_dir = home.join("inbox");
-    if !inbox_dir.exists() {
-        return;
-    }
-    let entries = match std::fs::read_dir(&inbox_dir) {
-        Ok(e) => e,
-        Err(_) => return,
-    };
-    let ts = chrono::Utc::now().format("%Y%m%dT%H%M%SZ").to_string();
-    let recovery_dir = home.join("inbox.recovery").join(&ts);
-    let mut recovered = 0u32;
-
-    for entry in entries.flatten() {
-        let path = entry.path();
-        let name = entry.file_name();
-        let name_str = name.to_string_lossy();
-
-        // Stale tmp files from interrupted atomic appends
-        if name_str.ends_with(".tmp") {
-            ensure_recovery_dir(&recovery_dir);
-            let dest = recovery_dir.join(&name);
-            if std::fs::rename(&path, &dest).is_ok() {
-                recovered += 1;
-            }
-            continue;
-        }
-
-        // Check JSONL files for corrupt trailing lines
-        if name_str.ends_with(".jsonl") {
-            if let Ok(content) = std::fs::read_to_string(&path) {
-                let lines: Vec<&str> = content.lines().collect();
-                let bad: Vec<&&str> = lines
-                    .iter()
-                    .filter(|l| {
-                        !l.trim().is_empty() && serde_json::from_str::<InboxMessage>(l).is_err()
-                    })
-                    .collect();
-                if !bad.is_empty() {
-                    // Move entire file to recovery, agent gets a fresh start
-                    ensure_recovery_dir(&recovery_dir);
-                    let dest = recovery_dir.join(&name);
-                    if std::fs::rename(&path, &dest).is_ok() {
-                        recovered += 1;
-                    }
-                }
-            }
-        }
-    }
-    if recovered > 0 {
-        tracing::warn!(
-            count = recovered,
-            dir = %recovery_dir.display(),
-            "inbox: recovered half-written files"
-        );
-    }
-}
-
-fn ensure_recovery_dir(dir: &Path) {
-    std::fs::create_dir_all(dir).ok();
-}
-
-/// Type-safe notification source — replaces raw string conventions.
-pub enum NotifySource<'a> {
-    /// Message from a channel user (Telegram, Discord, etc.).
-    Channel(&'a str, crate::channel::ChannelKind),
-    /// Message from another agent instance (e.g., "dev").
-    Agent(&'a str),
-    /// System message (e.g., "replace", "ci").
-    System(&'a str),
-}
-
-impl fmt::Display for NotifySource<'_> {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            Self::Channel(user, kind) => {
-                let kind_str = match kind {
-                    crate::channel::ChannelKind::Telegram => "telegram",
-                    crate::channel::ChannelKind::Discord => "discord",
-                };
-                write!(f, "user:{user} via {kind_str}")
-            }
-            Self::Agent(name) => write!(f, "from:{name}"),
-            Self::System(label) => write!(f, "system:{label}"),
-        }
-    }
-}
-
-impl NotifySource<'_> {
-    fn reply_hint(&self) -> Cow<'static, str> {
-        match self {
-            Self::Channel(_, _) => {
-                "\n(Reply using the reply tool — do NOT respond with direct text)".into()
-            }
-            Self::Agent(sender) => {
-                format!("\n(Reply using the send tool with target_instance=\"{sender}\")").into()
-            }
-            Self::System(_) => "".into(),
-        }
-    }
-}
-
-#[derive(Debug, Clone, Default, Serialize, Deserialize)]
-pub struct InboxMessage {
-    #[serde(default)]
-    pub schema_version: u32,
-    #[serde(default)]
-    pub id: Option<String>,
-    pub from: String,
-    pub text: String,
-    pub kind: Option<String>,
-    pub timestamp: String,
-    /// Channel source identity (typed). Additive field replacing the ad-hoc
-    /// `kind: "telegram"` misuse from telegram adapter.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub channel: Option<crate::channel::ChannelKind>,
-    #[serde(default)]
-    pub read_at: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub thread_id: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub parent_id: Option<String>,
-    /// How the message was delivered: "pty" (injected to agent PTY) or
-    /// "inbox_fallback" (daemon down, written to inbox file only).
-    /// Absent on legacy messages; backwards-compatible via `#[serde(default)]`.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub delivery_mode: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub task_id: Option<String>,
-    /// Force metadata — set when delegate_task used force=true (overrides busy gate).
-    /// Serde alias "interrupt_meta" for backwards-compat with Sprint 8-9 inbox JSONL.
-    #[serde(
-        default,
-        skip_serializing_if = "Option::is_none",
-        alias = "interrupt_meta"
-    )]
-    pub force_meta: Option<ForceMeta>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub correlation_id: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub reviewed_head: Option<String>,
-    /// Inbound media attachments (additive). Channel adapters download media
-    /// to a local path before enqueue. Empty for text-only messages.
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub attachments: Vec<crate::channel::event::Attachment>,
-    /// If the user replied to a specific bot message, this is that message's id.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub in_reply_to_msg_id: Option<String>,
-    /// Excerpt of the replied-to message (first 200 chars + author tag).
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub in_reply_to_excerpt: Option<String>,
-    /// ID of a newer message that supersedes this one (e.g. ci-watch SHA update).
-    /// Messages with superseded_by set are excluded from drain by default.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub superseded_by: Option<String>,
-    /// Sender's instance ID (UUIDv4) for audit trail. Display: name (id8).
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub from_id: Option<String>,
-    /// Sprint 54 layer-5 broadcast visibility: populated when this message
-    /// arrived via a `send` broadcast (team / targets / tags fan-out).
-    /// Absent on unicast — same conditional pattern as `attachments`. Lets
-    /// recipient agents distinguish "broadcast to N peers" from a direct
-    /// 1-on-1 message at JSON-metadata vantage; the PTY-inject header
-    /// surfaces the same context via `team=` / `broadcast=` fields.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub broadcast_context: Option<BroadcastContext>,
-    /// Dispatch schema fields (Issue #649, Phase 1)
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub sequencing: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub eta_minutes: Option<u32>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub reporting_cadence: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub worktree_binding_required: Option<bool>,
-    /// #1031: pull-request number associated with the message. Populated
-    /// by the ci_watch poller on `[ci-pass]` / `[ci-ready-for-action]`
-    /// events from the pr_state aggregator cache, so reviewers can post
-    /// §3.12 verdict mirrors via `gh pr comment <N>` without a
-    /// `gh pr list --head` lookup. Absent on legacy messages and
-    /// non-CI events — `#[serde(default)]` ensures backwards-compat
-    /// with pre-#1031 JSONL inboxes.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub pr_number: Option<u64>,
-}
-
-/// Metadata attached to a forced delegation (busy gate override).
-/// Renamed from InterruptMeta in Sprint 10 (semantic correction).
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ForceMeta {
-    #[serde(alias = "interrupted")]
-    pub forced: bool,
-    pub reason: String,
-    #[serde(alias = "interrupted_at")]
-    pub forced_at: String,
-}
-
-/// Sprint 54 layer-5 broadcast visibility: routing-time metadata captured at
-/// `handle_broadcast` and threaded through the SEND path so the recipient
-/// agent can tell broadcast from unicast. `team` is `Some` only for
-/// team-based broadcasts; `targets`/`count` are populated for every fan-out
-/// (including direct `targets=[…]` / `tags=[…]` modes). Routing behavior is
-/// unaffected — this struct is metadata-only.
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub struct BroadcastContext {
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub team: Option<String>,
-    #[serde(default)]
-    pub targets: Vec<String>,
-    #[serde(default)]
-    pub count: usize,
-}
-
-impl InboxMessage {
-    /// Latest schema version this binary can read and write.
-    pub const CURRENT_VERSION: u32 = 1;
-
-    pub fn new_system(
-        from: impl Into<String>,
-        kind: impl Into<String>,
-        text: impl Into<String>,
-    ) -> Self {
-        Self {
-            from: from.into(),
-            text: text.into(),
-            kind: Some(kind.into()),
-            timestamp: chrono::Utc::now().to_rfc3339(),
-            ..Default::default()
-        }
-    }
-
-    pub fn with_correlation_id(mut self, id: impl Into<String>) -> Self {
-        self.correlation_id = Some(id.into());
-        self
-    }
-
-    pub fn with_reviewed_head(mut self, sha: impl Into<String>) -> Self {
-        self.reviewed_head = Some(sha.into());
-        self
-    }
-
-    pub fn with_delivery_mode(mut self, mode: impl Into<String>) -> Self {
-        self.delivery_mode = Some(mode.into());
-        self
-    }
-}
-
-/// Status of a specific inbox message.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum MessageStatus {
-    /// Message was read at the given timestamp.
-    ReadAt(String, Option<String>), // (read_at, delivery_mode)
-    /// Message exists but has not been read and has expired (>30d).
-    UnreadExpired,
-    /// Message not found.
-    NotFound,
-}
-
-pub(crate) fn inbox_path(home: &Path, name: &str) -> PathBuf {
-    home.join("inbox").join(format!("{name}.jsonl"))
-}
-
-/// Sprint 46 P2: resolve inbox path by InstanceId when available.
-/// Migrates legacy name-based files to id-based on first access.
-pub(crate) fn inbox_path_resolved(home: &Path, name: &str) -> PathBuf {
-    // Only use id-based path when the instance has a real ID in fleet.yaml
-    // (backfilled by P1). Instances without an ID use name-based paths.
-    let id = crate::fleet::FleetConfig::load(&crate::fleet::fleet_yaml_path(home))
-        .ok()
-        .and_then(|c| {
-            c.instances
-                .get(name)
-                .and_then(|i| i.id.as_deref())
-                .and_then(crate::types::InstanceId::parse)
-        });
-    let Some(id) = id else {
-        return inbox_path(home, name);
-    };
-    let id_path = home.join("inbox").join(format!("{}.jsonl", id.full()));
-    if id_path.exists() {
-        return id_path;
-    }
-    let name_path = inbox_path(home, name);
-    if name_path.exists() {
-        // Migrate: create symlink from id-based to name-based (or copy on Windows)
-        if let Some(parent) = id_path.parent() {
-            let _ = std::fs::create_dir_all(parent);
-        }
-        #[cfg(unix)]
-        {
-            let _ = std::os::unix::fs::symlink(&name_path, &id_path);
-        }
-        #[cfg(windows)]
-        {
-            let _ = std::fs::copy(&name_path, &id_path);
-        }
-        return id_path;
-    }
-    // New instance — use id-based path directly
-    id_path
-}
-
-/// Acquire a per-agent flock and run `f` with the inbox path.
-/// All read-modify-write operations on an agent's inbox (enqueue, drain,
-/// sweep_expired) must go through this helper to prevent concurrent races.
-fn with_inbox_lock<T>(home: &Path, name: &str, f: impl FnOnce(&Path) -> T) -> anyhow::Result<T> {
-    let path = inbox_path_resolved(home, name);
-    if let Some(parent) = path.parent() {
-        std::fs::create_dir_all(parent)?;
-    }
-    let lock_path = path.with_extension("jsonl.lock");
-    let _lock = crate::store::acquire_file_lock(&lock_path)?;
-    Ok(f(&path))
-}
-
-/// Enqueue a message — atomic append via flock + tmp + fsync + rename.
-///
-/// Returns an error when the inbox is in readonly mode (disk full).
-/// Callers should invoke [`check_disk_space`] periodically (e.g. daemon tick);
-/// enqueue only reads the cached flag.
-///
-/// Concurrent safety: a per-agent flock via [`with_inbox_lock`] serialises
-/// all read-modify-write operations (enqueue, drain, sweep) on the same
-/// agent inbox (cross-process safe).
-pub fn enqueue(home: &Path, name: &str, mut msg: InboxMessage) -> anyhow::Result<()> {
-    if is_readonly() {
-        anyhow::bail!("inbox readonly: disk space critically low");
-    }
-    msg.schema_version = InboxMessage::CURRENT_VERSION;
-    ensure_msg_id(&mut msg);
-    let line = format!("{}\n", serde_json::to_string(&msg)?);
-
-    with_inbox_lock(home, name, |path| {
-        // H1: append-only write — O(1) instead of O(n) read-all+rewrite.
-        // The file is a JSONL append log; we only need to add one line.
-        let result = (|| -> anyhow::Result<()> {
-            let mut f = std::fs::OpenOptions::new()
-                .create(true)
-                .append(true)
-                .open(path)?;
-            f.write_all(line.as_bytes())?;
-            f.sync_all()?;
-            Ok(())
-        })();
-        result
-    })?
-}
-
-/// Assign a stable `msg.id` when absent. Shared between [`enqueue`] and
-/// [`enqueue_with_idle_hint`] so the latter can pre-stamp an id before
-/// the enqueue, then reference it in the PTY hint without consuming the
-/// message-by-value twice.
-fn ensure_msg_id(msg: &mut InboxMessage) {
-    if msg.id.is_some() {
-        return;
-    }
-    use std::sync::atomic::{AtomicU64, Ordering as AtOrd};
-    static MSG_SEQ: AtomicU64 = AtomicU64::new(0);
-    let ts = chrono::Utc::now().format("%Y%m%d%H%M%S%6f");
-    let seq = MSG_SEQ.fetch_add(1, AtOrd::Relaxed);
-    msg.id = Some(format!("m-{ts}-{seq}"));
-}
-
-/// Mark prior unread ci-watch messages for the same repo+branch as superseded.
-/// Called before enqueuing a new ci-watch notification so stale events don't surface.
-pub fn mark_ci_watch_superseded(
-    home: &Path,
-    instance: &str,
-    repo_branch_key: &str,
-    new_msg_id: &str,
-) {
-    let path = inbox_path_resolved(home, instance);
-    if !path.exists() {
-        return;
-    }
-    let _ = with_inbox_lock(home, instance, |path| -> anyhow::Result<()> {
-        let content = std::fs::read_to_string(path).unwrap_or_default();
-        let mut changed = false;
-        let mut lines: Vec<String> = Vec::new();
-        for line in content.lines() {
-            if line.trim().is_empty() {
-                lines.push(line.to_string());
-                continue;
-            }
-            if let Ok(mut msg) = serde_json::from_str::<InboxMessage>(line) {
-                if msg.read_at.is_none()
-                    && msg.superseded_by.is_none()
-                    && msg.kind.as_deref() == Some("ci-watch")
-                    && msg.from == "system:ci"
-                    && msg.text.contains(repo_branch_key)
-                {
-                    msg.superseded_by = Some(new_msg_id.to_string());
-                    changed = true;
-                }
-                lines.push(serde_json::to_string(&msg).unwrap_or_else(|_| line.to_string()));
-            } else {
-                lines.push(line.to_string());
-            }
-        }
-        if changed {
-            let tmp = path.with_extension("jsonl.tmp");
-            let mut f = std::fs::OpenOptions::new()
-                .create(true)
-                .write(true)
-                .truncate(true)
-                .open(&tmp)?;
-            use std::io::Write;
-            for l in &lines {
-                writeln!(f, "{l}")?;
-            }
-            f.sync_all()?;
-            std::fs::rename(&tmp, path)?;
-        }
-        Ok(())
-    });
-}
-
-/// Drain unread messages: mark them with `read_at` and write back.
-/// Returns only the messages that were previously unread.
-///
-/// Soft-delete semantics: messages stay in the JSONL file with `read_at`
-/// set; [`sweep_expired`] removes them later based on TTL rules.
-/// Uses atomic tmp+fsync+rename for crash safety.
-pub fn drain(home: &Path, name: &str) -> Vec<InboxMessage> {
-    let path = inbox_path_resolved(home, name);
-
-    if !path.exists() && !path.with_extension("draining").exists() {
-        return Vec::new();
-    }
-
-    match with_inbox_lock(home, name, |path| {
-        // Leftover from a crashed predecessor — consume it first, under lock
-        // to prevent duplicate delivery when two drain callers race.
-        let tmp = path.with_extension("draining");
-        if tmp.exists() {
-            return read_drain_file(&tmp);
-        }
-
-        let content = match std::fs::read_to_string(path) {
-            Ok(c) => c,
-            Err(_) => return Vec::new(),
-        };
-
-        let now = chrono::Utc::now().to_rfc3339();
-        let mut unread = Vec::new();
-        let mut all_messages: Vec<InboxMessage> = Vec::new();
-
-        for line in content.lines() {
-            if line.trim().is_empty() {
-                continue;
-            }
-            let mut msg: InboxMessage = match serde_json::from_str(line) {
-                Ok(m) => m,
-                Err(_) => continue,
-            };
-            if msg.schema_version > InboxMessage::CURRENT_VERSION {
-                tracing::error!(
-                    found = msg.schema_version,
-                    supported = InboxMessage::CURRENT_VERSION,
-                    "dropping inbox message written by newer schema version"
-                );
-                continue;
-            }
-            if msg.read_at.is_none() {
-                // M6: skip superseded messages — they are stale
-                if msg.superseded_by.is_some() {
-                    msg.read_at = Some(now.clone());
-                    all_messages.push(msg);
-                    continue;
-                }
-                msg.read_at = Some(now.clone());
-                // #836: signal post-consume so the notification-dedup
-                // ledger can suppress redundant re-inject attempts
-                // from the rate-limit retry path. The ledger only has
-                // a record when notify_agent earlier called
-                // `record_inject(agent, msg_id)`; for messages whose
-                // delivery never went through the PTY-inject path
-                // (inbox-only or pre-#836 flight), `mark_consumed` is
-                // a no-op.
-                if let Some(ref id) = msg.id {
-                    crate::daemon::notification_dedup::global().mark_consumed(name, id);
-                }
-                unread.push(msg.clone());
-            }
-            all_messages.push(msg);
-        }
-
-        // Sprint 52: set reply_to on dequeue — last channel-sourced message wins.
-        if let Some(channel_msg) = unread.iter().rev().find(|m| m.channel.is_some()) {
-            let channel_name = match channel_msg.channel.as_ref().expect("checked") {
-                crate::channel::ChannelKind::Telegram => "telegram",
-                crate::channel::ChannelKind::Discord => "discord",
-            };
-            crate::daemon::heartbeat_pair::update_with(name, |p| {
-                p.reply_to_channel = Some(channel_name.to_string());
-                p.reply_to_input_id = Some(p.reply_to_input_id.unwrap_or(0) + 1);
-                p.reply_to_set_at_ms = crate::daemon::heartbeat_pair::now_ms() as i64;
-                p.mirror_dispatched_for_turn = false;
-                p.mirror_skip_until_next_turn = false;
-            });
-        }
-
-        if !unread.is_empty() {
-            let write_tmp = path.with_extension("jsonl.tmp");
-            let result = (|| -> anyhow::Result<()> {
-                let mut f = std::fs::OpenOptions::new()
-                    .create(true)
-                    .write(true)
-                    .truncate(true)
-                    .open(&write_tmp)?;
-                for m in &all_messages {
-                    writeln!(f, "{}", serde_json::to_string(m)?)?;
-                }
-                f.sync_all()?;
-                std::fs::rename(&write_tmp, path)?;
-                Ok(())
-            })();
-            if let Err(e) = result {
-                tracing::warn!(error = %e, "inbox drain write-back failed");
-            }
-        }
-
-        unread
-    }) {
-        Ok(msgs) => msgs,
-        Err(e) => {
-            tracing::warn!(error = %e, "inbox drain lock failed");
-            Vec::new()
-        }
-    }
-}
-
-fn read_drain_file(tmp: &Path) -> Vec<InboxMessage> {
-    let content = match std::fs::read_to_string(tmp) {
-        Ok(c) => c,
-        // Leave `.draining` in place so the next drain call retries; the
-        // previous implementation early-returned without removing, but also
-        // returned empty even on success when read_to_string returned Err
-        // after the earlier remove had run — which was impossible to recover.
-        Err(e) => {
-            tracing::warn!(
-                path = %tmp.display(),
-                error = %e,
-                "inbox drain read failed; .draining retained for retry"
-            );
-            return Vec::new();
-        }
-    };
-    let messages: Vec<InboxMessage> = content
-        .lines()
-        .filter_map(|l| {
-            let msg: InboxMessage = serde_json::from_str(l).ok()?;
-            if msg.schema_version > InboxMessage::CURRENT_VERSION {
-                tracing::error!(
-                    found = msg.schema_version,
-                    supported = InboxMessage::CURRENT_VERSION,
-                    "dropping inbox message written by newer schema version"
-                );
-                return None;
-            }
-            Some(msg)
-        })
-        .collect();
-    // Remove only AFTER a successful read+parse so crashes between read and
-    // remove still leave the data on disk for the next drain to recover.
-    if let Err(e) = std::fs::remove_file(tmp) {
-        tracing::warn!(path = %tmp.display(), error = %e, "inbox drain cleanup failed");
-    }
-    messages
-}
-
-// INLINE_THRESHOLD removed — deliver() now always enqueues regardless of length.
-
-/// Count unread messages (read_at == None) for an agent.
-pub fn unread_count(home: &Path, name: &str) -> (usize, Option<chrono::DateTime<chrono::Utc>>) {
-    let path = inbox_path_resolved(home, name);
-    let content = match std::fs::read_to_string(&path) {
-        Ok(c) => c,
-        Err(_) => return (0, None),
-    };
-    let mut count = 0usize;
-    let mut oldest: Option<chrono::DateTime<chrono::Utc>> = None;
-    for line in content.lines() {
-        if let Ok(msg) = serde_json::from_str::<InboxMessage>(line) {
-            if msg.read_at.is_none() {
-                count += 1;
-                if let Ok(ts) = chrono::DateTime::parse_from_rfc3339(&msg.timestamp) {
-                    let ts_utc = ts.with_timezone(&chrono::Utc);
-                    if oldest.is_none_or(|t| t > ts_utc) {
-                        oldest = Some(ts_utc);
-                    }
-                }
-            }
-        }
-    }
-    (count, oldest)
-}
-
-/// Size threshold for header-only PTY injection. Messages with body > this
-/// value inject only a structured header line; the full body stays in inbox.
-// #1065 removed the only production caller (`handle_send`'s `inject_msg`
-// build); test diagnostics + format-header callers still consume it.
-#[allow(dead_code)]
-pub const HEADER_SIZE_THRESHOLD: usize = 300;
-
-/// Returns true when the `AGEND_POINTER_ONLY_INJECT` feature flag is set to "1".
-/// When enabled, PTY injection uses header-only format for all messages,
-/// forcing agents to call `inbox` to read content (solves dispatch non-FIFO).
-pub fn pointer_only_inject() -> bool {
-    // G3 H1: read DaemonConfig instead of env var (thread-safe)
-    crate::daemon_config::get().pointer_only_inject
-}
-
-/// ANSI-colored header prefix for visual distinction in terminal.
-pub const HEADER_PREFIX: &str = "\x1b[44;97m[AGEND-MSG]\x1b[0m";
-
-/// #982: ANSI-colored prefix for idle-hint headers emitted alongside
-/// daemon-side inbox enqueues. Distinct from [`HEADER_PREFIX`] so
-/// recipients can distinguish the canonical message body inject from
-/// the lightweight `(use inbox tool)` pointer signalling a new
-/// pending entry. Background is yellow-on-black for contrast.
-pub const PENDING_HEADER_PREFIX: &str = "\x1b[43;30m[AGEND-MSG-PENDING]\x1b[0m";
-
-/// Plain-text system message prefix (without ANSI colors).
-/// Used for detection/matching in agent PTY output parsing.
-pub const SYSTEM_MSG_PREFIX: &str = "[AGEND-MSG]";
-
-/// Agent-to-agent message prefix.
-pub const AGENT_MSG_PREFIX: &str = "[from:";
-
-/// Sanitize a value for header inclusion: replace control chars with space.
-fn sanitize_header_value(s: &str) -> String {
-    s.chars()
-        .map(|c| if c.is_control() { ' ' } else { c })
-        .collect()
-}
-
-/// Format a single-line structured header for PTY injection.
-/// Fields: from / id / kind / thread / parent / size.
-/// Optional fields (thread/parent) omitted when None.
-///
-/// #1065 removed the production call site (`handle_send` now routes
-/// through `enqueue_with_idle_hint` which builds its own
-/// `[AGEND-MSG-PENDING]` hint). Retained `pub` for test diagnostics
-/// (e.g. `src/instructions.rs` instruction-text invariant checks).
-#[allow(dead_code)]
-pub fn format_header(msg: &InboxMessage) -> String {
-    // #761: strip the redundant `from:` prefix that `Source::Agent`'s
-    // Display impl (inbox.rs:144) adds. `strip_prefix` returns `Some`
-    // for agent sources and `None` otherwise, so `system:` / `user:`
-    // namespaces survive untouched via the `unwrap_or` fallback.
-    let from_value = msg.from.strip_prefix("from:").unwrap_or(&msg.from);
-    let mut parts = vec![
-        HEADER_PREFIX.to_string(),
-        format!("from={}", sanitize_header_value(from_value)),
-    ];
-    if let Some(ref id) = msg.id {
-        parts.push(format!("id={}", sanitize_header_value(id)));
-    }
-    if let Some(ref kind) = msg.kind {
-        parts.push(format!("kind={}", sanitize_header_value(kind)));
-    }
-    if let Some(ref thread) = msg.thread_id {
-        parts.push(format!("thread={}", sanitize_header_value(thread)));
-    }
-    if let Some(ref parent) = msg.parent_id {
-        parts.push(format!("parent={}", sanitize_header_value(parent)));
-    }
-    parts.push(format!("size={}", msg.text.chars().count()));
-    if !msg.attachments.is_empty() {
-        let paths: Vec<&str> = msg
-            .attachments
-            .iter()
-            .map(|a| a.path.to_str().unwrap_or("?"))
-            .collect();
-        parts.push(format!("attachments=[{}]", paths.join(",")));
-    }
-    // Sprint 54 layer-5 broadcast visibility: surface broadcast routing in
-    // the PTY-inject header so the recipient agent can tell broadcast from
-    // unicast at first glance. `broadcast=N` always emits when context is
-    // present; `team=NAME` only when team-based fan-out (absent for direct
-    // `targets=[…]` / `tags=[…]` modes per qa-test smoke success criteria).
-    if let Some(ref ctx) = msg.broadcast_context {
-        parts.push(format!("broadcast={}", ctx.count));
-        if let Some(ref team) = ctx.team {
-            parts.push(format!("team={}", sanitize_header_value(team)));
-        }
-    }
-    if let Some(ref excerpt) = msg.in_reply_to_excerpt {
-        parts.push(format!(
-            "reply_to_excerpt={}",
-            sanitize_header_value(excerpt)
-        ));
-    }
-    parts.join(" ")
-}
-
-/// Build the `in_reply_to_excerpt` value for inbound replies.
-///
-/// Returns `None` when `text` is empty so callers can `and_then` over an
-/// optional reply target. Otherwise truncates to 200 chars (Unicode-safe via
-/// `chars().take(...)`) and appends `…` if the original exceeded 200 chars.
-pub(crate) fn build_excerpt(text: &str, author: &str) -> Option<String> {
-    if text.is_empty() {
-        return None;
-    }
-    let truncated: String = text.chars().take(200).collect();
-    let ellipsis = if text.chars().count() > 200 {
-        "…"
-    } else {
-        ""
-    };
-    Some(format!("[{author}] {truncated}{ellipsis}"))
-}
-
-/// Format a single-line event header (non-message events like poll-reminder).
-/// Uses the same ANSI prefix as [`format_header`] for visual consistency.
-pub fn format_event_header(kind: &str, fields: &[(&str, &str)]) -> String {
-    let mut parts = vec![
-        HEADER_PREFIX.to_string(),
-        format!("kind={}", sanitize_header_value(kind)),
-    ];
-    for (k, v) in fields {
-        parts.push(format!("{}={}", k, sanitize_header_value(v)));
-    }
-    parts.join(" ")
-}
-
-/// Sweep expired messages from all inbox files.
-/// - read_at.is_some() && elapsed > 7 days → delete
-/// - read_at.is_none() && elapsed > 30 days → delete
-pub fn sweep_expired(home: &Path) {
-    let inbox_dir = home.join("inbox");
-    let entries = match std::fs::read_dir(&inbox_dir) {
-        Ok(e) => e,
-        Err(_) => return,
-    };
-    let now = chrono::Utc::now();
-    for entry in entries.flatten() {
-        let path = entry.path();
-        if path.extension().and_then(|e| e.to_str()) != Some("jsonl") {
-            continue;
-        }
-        // Extract agent name from filename (e.g. "agent1.jsonl" → "agent1")
-        let agent_name = match path.file_stem().and_then(|s| s.to_str()) {
-            Some(n) => n.to_string(),
-            None => continue,
-        };
-        let _ = with_inbox_lock(home, &agent_name, |path| {
-            let content = match std::fs::read_to_string(path) {
-                Ok(c) => c,
-                Err(_) => return,
-            };
-            let mut kept: Vec<String> = Vec::new();
-            let mut changed = false;
-            for line in content.lines() {
-                if line.trim().is_empty() {
-                    continue;
-                }
-                let msg: InboxMessage = match serde_json::from_str(line) {
-                    Ok(m) => m,
-                    Err(_) => continue,
-                };
-                let ts = chrono::DateTime::parse_from_rfc3339(&msg.timestamp)
-                    .map(|dt| dt.with_timezone(&chrono::Utc))
-                    .unwrap_or(now);
-                let age = now.signed_duration_since(ts);
-                let expired = match &msg.read_at {
-                    Some(_) => age > chrono::Duration::days(7),
-                    None => age > chrono::Duration::days(30),
-                };
-                if expired {
-                    changed = true;
-                } else {
-                    kept.push(line.to_string());
-                }
-            }
-            if changed {
-                if kept.is_empty() {
-                    let _ = std::fs::remove_file(path);
-                } else {
-                    let tmp = path.with_extension("jsonl.tmp");
-                    let result = (|| -> anyhow::Result<()> {
-                        let mut f = std::fs::OpenOptions::new()
-                            .create(true)
-                            .write(true)
-                            .truncate(true)
-                            .open(&tmp)?;
-                        for l in &kept {
-                            writeln!(f, "{l}")?;
-                        }
-                        f.sync_all()?;
-                        std::fs::rename(&tmp, path)?;
-                        Ok(())
-                    })();
-                    if let Err(e) = result {
-                        tracing::warn!(error = %e, "inbox sweep write-back failed");
-                    }
-                }
-            }
-        });
-    }
-}
-
-/// Look up a message by ID in a specific agent's inbox file.
-/// If `instance` is provided, only that agent's inbox is searched.
-pub fn describe_message(home: &Path, msg_id: &str, instance: &str) -> MessageStatus {
-    let path = inbox_path_resolved(home, instance);
-    if !path.exists() {
-        return MessageStatus::NotFound;
-    }
-    let content = match std::fs::read_to_string(&path) {
-        Ok(c) => c,
-        Err(_) => return MessageStatus::NotFound,
-    };
-    let now = chrono::Utc::now();
-    for line in content.lines() {
-        let msg: InboxMessage = match serde_json::from_str(line) {
-            Ok(m) => m,
-            Err(_) => continue,
-        };
-        if msg.id.as_deref() != Some(msg_id) {
-            continue;
-        }
-        if let Some(ref read_at) = msg.read_at {
-            return MessageStatus::ReadAt(read_at.clone(), msg.delivery_mode.clone());
-        }
-        let ts = chrono::DateTime::parse_from_rfc3339(&msg.timestamp)
-            .map(|dt| dt.with_timezone(&chrono::Utc))
-            .unwrap_or(now);
-        if now.signed_duration_since(ts) > chrono::Duration::days(30) {
-            return MessageStatus::UnreadExpired;
-        }
-        return MessageStatus::NotFound;
-    }
-    MessageStatus::NotFound
-}
-
-/// Get all messages in a thread, ordered by timestamp.
-/// If `instance` is Some, only scan that agent's inbox; otherwise scan all.
-pub fn get_thread(home: &Path, thread_id: &str, instance: Option<&str>) -> Vec<InboxMessage> {
-    let inbox_dir = home.join("inbox");
-    let entries = match std::fs::read_dir(&inbox_dir) {
-        Ok(e) => e,
-        Err(_) => return Vec::new(),
-    };
-    let mut msgs = Vec::new();
-    for entry in entries.flatten() {
-        let path = entry.path();
-        if path.extension().and_then(|e| e.to_str()) != Some("jsonl") {
-            continue;
-        }
-        if let Some(inst) = instance {
-            let stem = path.file_stem().and_then(|s| s.to_str()).unwrap_or("");
-            if stem != inst {
-                continue;
-            }
-        }
-        let content = match std::fs::read_to_string(&path) {
-            Ok(c) => c,
-            Err(_) => continue,
-        };
-        for line in content.lines() {
-            if let Ok(msg) = serde_json::from_str::<InboxMessage>(line) {
-                if msg.thread_id.as_deref() == Some(thread_id) {
-                    msgs.push(msg);
-                }
-            }
-        }
-    }
-    msgs.sort_by(|a, b| a.timestamp.cmp(&b.timestamp));
-    msgs
-}
-
-/// Look up a message by ID across all inbox files. Returns the message if found.
-pub fn find_message(home: &Path, msg_id: &str) -> Option<InboxMessage> {
-    let inbox_dir = home.join("inbox");
-    let entries = std::fs::read_dir(&inbox_dir).ok()?;
-    for entry in entries.flatten() {
-        let path = entry.path();
-        if path.extension().and_then(|e| e.to_str()) != Some("jsonl") {
-            continue;
-        }
-        let content = std::fs::read_to_string(&path).ok()?;
-        for line in content.lines() {
-            if let Ok(msg) = serde_json::from_str::<InboxMessage>(line) {
-                if msg.id.as_deref() == Some(msg_id) {
-                    return Some(msg);
-                }
-            }
-        }
-    }
-    None
-}
-
-/// Deliver a message: always enqueue to inbox JSONL for persistence,
-/// then inject to PTY (inline or pointer-only depending on feature flag).
-///
-/// Sprint 54 layer-5: `broadcast_context` is `Some` only when the call
-/// originates from a `handle_broadcast` per-target fan-out via the
-/// `agent_ops::send_to` daemon-unavailable fallback. The daemon-available
-/// path goes through `api::handlers::messaging::handle_send` and
-/// constructs its own `InboxMessage` with the same field.
-pub fn deliver(
-    home: &Path,
-    agent_name: &str,
-    source: &NotifySource<'_>,
-    text: &str,
-    _submit_key: &str,
-    kind: Option<String>,
-    broadcast_context: Option<BroadcastContext>,
-) {
-    let msg = InboxMessage {
-        from: source.to_string(),
-        text: text.to_string(),
-        kind,
-        timestamp: chrono::Utc::now().to_rfc3339(),
-        broadcast_context,
-        ..Default::default()
-    };
-    let _ = enqueue(home, agent_name, msg);
-    notify_agent(home, agent_name, source, text);
-}
-
-/// Sprint 54 silent-drop layer-4 hotfix: aggregate kind counts for the
-/// PTY-inject header `attachments=[…]` field. Returns `None` for empty
-/// so callers can `if let Some(s) = …` without producing trailing
-/// whitespace on no-attachment notifications.
-///
-/// Output format mirrors operator's `m-9` request: `1 photo, 2 document`
-/// — kind labels in lowercase, count-aggregated, comma-joined in a
-/// stable order so test fixtures don't depend on HashMap iteration.
-pub(crate) fn summarize_attachments_for_header(
-    attachments: &[crate::channel::event::Attachment],
-) -> Option<String> {
-    if attachments.is_empty() {
-        return None;
-    }
-    use crate::channel::event::AttachmentKind;
-    let mut counts: [usize; 5] = [0; 5];
-    for a in attachments {
-        let i = match a.kind {
-            AttachmentKind::Photo => 0,
-            AttachmentKind::Voice => 1,
-            AttachmentKind::Document => 2,
-            AttachmentKind::Video => 3,
-            AttachmentKind::Sticker => 4,
-        };
-        counts[i] += 1;
-    }
-    let labels = ["photo", "voice", "document", "video", "sticker"];
-    let parts: Vec<String> = counts
-        .iter()
-        .zip(labels.iter())
-        .filter(|(n, _)| **n > 0)
-        .map(|(n, label)| format!("{n} {label}"))
-        .collect();
-    Some(parts.join(", "))
-}
-
-/// Sprint 54 silent-drop layer-4 hotfix: human-readable body
-/// placeholder used when the inbox message has no text but attachments
-/// are present. Without this, `format_notification_for_inject` in
-/// inline mode produced an empty body — agent saw a content-less
-/// notification it couldn't action on.
-///
-/// Single attachment with `original_filename` → `[1 photo: cat.jpg]`.
-/// Single attachment without filename → `[1 photo attached]`.
-/// Multiple attachments → `[1 photo, 2 document attached]`.
-pub(crate) fn attachment_body_placeholder(
-    attachments: &[crate::channel::event::Attachment],
-) -> String {
-    if attachments.is_empty() {
-        return String::new();
-    }
-    let summary = summarize_attachments_for_header(attachments).unwrap_or_default();
-    if attachments.len() == 1 {
-        if let Some(name) = &attachments[0].original_filename {
-            return format!("[{summary}: {name}]");
-        }
-    }
-    format!("[{summary} attached]")
-}
-
-/// Pure function: build the notification string that will be injected to PTY.
-/// `pointer_only=true` → header-only (no body); `false` → inline text.
-///
-/// Sprint 54 silent-drop layer-4 hotfix: when `attachments` is non-empty,
-/// the header gains an `attachments=[…]` field and (in inline mode)
-/// empty-text notifications get a human-readable body placeholder
-/// instead of producing a content-less PTY inject. `attachments=[]`
-/// preserves pre-r0 behavior so non-Telegram callers stay unaffected.
-pub fn format_notification_for_inject(
-    pointer_only: bool,
-    source: &NotifySource<'_>,
-    text: &str,
-    attachments: &[crate::channel::event::Attachment],
-) -> String {
-    let attach_summary = summarize_attachments_for_header(attachments);
-    if pointer_only {
-        let mut header = format!(
-            "[{source}] [AGEND-MSG] size={} (use inbox tool)",
-            text.len()
-        );
-        if let Some(s) = &attach_summary {
-            header.push_str(&format!(" attachments=[{s}]"));
-        }
-        header.push_str(&source.reply_hint());
-        header
-    } else {
-        // Body-replace path: empty text + non-empty attachments would
-        // produce a content-less inline notification pre-r0. Substitute
-        // the placeholder so the agent at minimum sees what kind of
-        // media is waiting.
-        let display_text = if text.is_empty() && !attachments.is_empty() {
-            attachment_body_placeholder(attachments)
-        } else if text.chars().count() > 200 {
-            let truncated: String = text.chars().take(200).collect();
-            format!("{truncated}... (use the inbox MCP tool to read full message)")
-        } else {
-            text.to_string()
-        };
-        format!("[{source}] {display_text}{}", source.reply_hint())
-    }
-}
-
-pub fn notify_agent(home: &Path, agent_name: &str, source: &NotifySource<'_>, text: &str) {
-    notify_agent_with_attachments(home, agent_name, source, text, &[]);
-}
-
-/// Sprint 54 silent-drop layer-4 hotfix: variant of `notify_agent` that
-/// carries attachment metadata into the PTY-inject formatter. Callers
-/// that have an `InboxMessage` with attachments (currently just the
-/// telegram inbound path) use this so the agent sees the canonical
-/// `attachments=[…]` header field + a human-readable body placeholder
-/// when the inbox text is empty. Other callers (`agent_ops::send`,
-/// supervisor notifications) stick with the plain `notify_agent` —
-/// their text-only wrapper passes `&[]` here.
-pub fn notify_agent_with_attachments(
-    home: &Path,
-    agent_name: &str,
-    source: &NotifySource<'_>,
-    text: &str,
-    attachments: &[crate::channel::event::Attachment],
-) {
-    // Sprint 24 P1 (F-NEW-DAEMON-HEALTH-CLASSIFIER-1): record this
-    // central inject point so the daemon health classifier can
-    // distinguish "idle waiting (no input pending)" from "hung
-    // unresponsive (input pending past last response)". This hook
-    // intentionally lives here — `notify_agent` is the single PTY-inject
-    // entry consumed by every input source (channel notify, agent send,
-    // inbox compose-aware inject). No-daemon-context callers (tests
-    // wiring `notify_agent` outside daemon mode) still pay a cheap
-    // mutex acquire but don't observe behavioural change.
-    crate::daemon::heartbeat_pair::update_with(agent_name, |p| {
-        p.last_input_at_ms = crate::daemon::heartbeat_pair::now_ms();
-    });
-    let notification =
-        format_notification_for_inject(pointer_only_inject(), source, text, attachments);
-    compose_aware_inject(home, agent_name, &notification);
-    // #836: record the (agent, msg_id) tuple in the dedup ledger so a
-    // subsequent post-consume retry tick can suppress redundant
-    // re-inject attempts. We parse the msg_id back out of the
-    // formatted header (single source of truth — `format_header`
-    // already includes `id=<msg_id>` when set). Headers without an
-    // id field (event-style notifications, free-form acks) skip the
-    // record cleanly per the parser's `Option<String>` contract.
-    if let Some(msg_id) =
-        crate::daemon::notification_dedup::extract_msg_id_from_header(&notification)
-    {
-        crate::daemon::notification_dedup::global().record_inject(agent_name, &msg_id);
-    }
-    // Store raw body AFTER inject — overwrites any formatted text the inject
-    // handler may have stored. Ensures retry re-injects raw body, not header.
-    crate::daemon::heartbeat_pair::update_with(agent_name, |p| {
-        p.last_input_text = Some(text.to_string());
-    });
-}
-
-/// Compose-aware notification delivery: checks `is_composing` and enqueues
-/// if the target agent is mid-typing, otherwise injects **with** submit_key
-/// so idle agents actually wake up on the incoming notification. The
-/// is_composing guard (3 s input-idle window) preserves PR #81's race fix —
-/// user keystrokes never collide with a background submit — while restoring
-/// the pre-#81 behavior of actually delivering Telegram / notify_agent
-/// traffic to backends that don't poll inbox on their own.
-pub fn compose_aware_inject(home: &Path, agent_name: &str, notification: &str) {
-    // #911 dedup gate: (A)+(B) hybrid against the canonical
-    // `[AGEND-MSG]` header. Returns early when the msg was already
-    // consumed by the recipient — closes the operator-observable
-    // "header arrives after inbox drain" symptom.
-    //
-    // Mitigation-first: this is symptom containment + observability,
-    // not root-cause attribution. The producer paths that fire stale
-    // notify_agent calls are not (yet) identified; daemon-side log
-    // instrumentation is filed as a separate follow-up.
-    if should_suppress_911_reinject_with_ledger(
-        home,
-        agent_name,
-        notification,
-        crate::daemon::notification_dedup::global(),
-    ) {
-        return;
-    }
-    let _ = route_notification(home, agent_name, notification, |msg| {
-        inject_with_submit(home, agent_name, msg)
-    });
-}
-
-/// #911 dedup gate predicate. Hybrid (A)+(B) suppression check for
-/// a candidate PTY notification:
-///
-/// - **Fast path (B)** — `notification_dedup` ledger. If the
-///   `(agent, msg_id)` entry exists AND is flagged consumed AND
-///   within the 10-min TTL, suppress.
-/// - **Fallback path (A)** — when the ledger has no entry (TTL
-///   evicted at 10 min OR never recorded), read the agent's inbox
-///   JSONL for `msg_id` and check `read_at`. If `Some(_)`, the msg
-///   was drained — suppress.
-///
-/// Returns `true` to suppress, `false` to allow inject through.
-///
-/// Scope predicate: only canonical `[AGEND-MSG]` headers with an
-/// `id=` token. Free-form text and event-style headers (no `id=`)
-/// always return `false` — they aren't message-bound.
-///
-/// The two-path design closes H6 (cross-audit `m-20260518162253458320-159`):
-/// stale re-injects > 10 min post-drain miss the ledger but still get
-/// caught by the JSONL source-of-truth.
-///
-/// Production callers (compose_aware_inject) pass
-/// `notification_dedup::global()`. Tests construct local
-/// `Ledger::default()` instances for isolation.
-pub(crate) fn should_suppress_911_reinject_with_ledger(
-    home: &Path,
-    agent_name: &str,
-    notification: &str,
-    ledger: &crate::daemon::notification_dedup::Ledger,
-) -> bool {
-    // Reviewer condition C: only canonical `HEADER_PREFIX` headers
-    // gate (the ANSI-wrapped `[AGEND-MSG]` produced by `format_header`
-    // and `format_event_header`). Stray `id=` in free-form chat
-    // content with no ANSI prefix must NOT trigger.
-    if !notification.starts_with(HEADER_PREFIX) {
-        return false;
-    }
-    let Some(msg_id) = crate::daemon::notification_dedup::extract_msg_id_from_header(notification)
-    else {
-        return false;
-    };
-    // Fast path (B): in-memory ledger.
-    if ledger.should_suppress_reinject(agent_name, &msg_id) {
-        tracing::info!(
-            agent = %agent_name,
-            msg_id = %msg_id,
-            "#911 compose_aware_inject suppressed: dedup ledger hit"
-        );
-        return true;
-    }
-    // Fallback (A): JSONL read_at source-of-truth.
-    if msg_already_drained_in_jsonl(home, agent_name, &msg_id) {
-        tracing::info!(
-            agent = %agent_name,
-            msg_id = %msg_id,
-            "#911 compose_aware_inject suppressed: JSONL fallback (ledger MISS, read_at set)"
-        );
-        return true;
-    }
-    false
-}
-
-/// #982 B-narrow: scan `agent_name`'s inbox for a previously-drained
-/// dispatch (`kind ∈ {query, task}`, `read_at.is_some()`) that shares
-/// the given `correlation_id`. Used by `api::handlers::messaging` to
-/// override codex ack-absorption when an inbound `kind=report|update`
-/// is the reply to a blocking dispatch the recipient already consumed.
-///
-/// Tight predicate (positive override path):
-/// - kind matches `query` or `task` only — informational kinds never
-///   gate ack absorption
-/// - `read_at` must be set — an undrained dispatch means the recipient
-///   has not yet seen the parent so the inbound is not a blocking reply
-/// - `correlation_id` must match — same correlation chain
-///
-/// Best-effort: IO/parse failures default to `false` (preserve existing
-/// absorption behavior on file errors).
-pub fn has_drained_blocker_for_correlation(
-    home: &Path,
-    agent_name: &str,
-    correlation_id: &str,
-) -> bool {
-    let path = inbox_path_resolved(home, agent_name);
-    let Ok(content) = std::fs::read_to_string(&path) else {
-        return false;
-    };
-    content.lines().any(|line| {
-        let Ok(msg) = serde_json::from_str::<InboxMessage>(line) else {
-            return false;
-        };
-        msg.correlation_id.as_deref() == Some(correlation_id)
-            && msg.read_at.is_some()
-            && matches!(msg.kind.as_deref(), Some("query") | Some("task"))
-    })
-}
-
-/// Read the agent's inbox JSONL and return `true` iff a message with
-/// the given `msg_id` exists AND has `read_at` set. Best-effort: any
-/// IO/parse error returns `false` so the caller defaults to allowing
-/// the inject through (better operator-visible spurious header than
-/// silently dropped legit retry).
-fn msg_already_drained_in_jsonl(home: &Path, agent_name: &str, msg_id: &str) -> bool {
-    let path = inbox_path(home, agent_name);
-    let Ok(content) = std::fs::read_to_string(&path) else {
-        return false;
-    };
-    content.lines().any(|line| {
-        let Ok(msg) = serde_json::from_str::<InboxMessage>(line) else {
-            return false;
-        };
-        msg.id.as_deref() == Some(msg_id) && msg.read_at.is_some()
-    })
-}
-
-// #1065: `compose_aware_send` removed — `handle_send` now routes through
-// `enqueue_with_idle_hint` for the PTY path, unifying the wake signal with
-// daemon-emitted auto-wake. The composing-aware injection primitive is now
-// solely `compose_aware_inject` (used by `enqueue_with_idle_hint`).
-
-/// #982 Direction A: persist a message to the recipient's inbox AND emit a
-/// best-effort `[AGEND-MSG-PENDING]` PTY hint so daemon-side events do not
-/// strand silently when the recipient is at an idle prompt.
-///
-/// Mirror of [`enqueue`] for daemon scanner sites (waiting_on_stale,
-/// dispatch_idle, pr_state, etc.) that own no inter-agent send path of
-/// their own. The hint reuses [`compose_aware_inject`] so `is_composing`
-/// deferral, `notification_queue` flushing, and #911 dedup behave
-/// identically to the channel/notify_agent path.
-///
-/// # Do NOT migrate
-///
-/// The following sites MUST keep calling [`enqueue`] directly:
-///
-/// - `agent_ops::fallback_deliver` — silent-degrade path for the inter-
-///   agent send fallback. It already invokes
-///   [`crate::inbox::notify_agent`] (which routes through
-///   [`compose_aware_inject`]) on the same hop, so migrating would
-///   double-inject and risk re-entry into the API loopback.
-/// - `api::handlers::messaging::handle_send` — (post-#1065) MIGRATED to
-///   this call. Pre-#1065 used bare `enqueue` + `compose_aware_send`
-///   which injected the full `[AGEND-MSG]` header; the size pressure
-///   extended the inject window past the submit-key delay window
-///   (race-condition on `\r`). Migrating here closes the divergence
-///   with daemon-emitted auto-wake.
-/// - `channel::telegram::inbound` status-keyword path — paired with
-///   [`crate::inbox::notify_agent`] on the same hop.
-/// - `daemon::supervisor` member-state-change emit — paired with
-///   [`crate::inbox::notify_agent`].
-/// - `daemon::ci_watch::poller`: historically held three raw `enqueue`
-///   sites. #1030 migrated the [ci-pass] subscriber fan-out (line
-///   ~1092) — that one was paired with the companion
-///   [`crate::agent::inject_to_agent`] call at ~1040 but the inject
-///   only wakes attentive agents; idle codex / claude-code backends
-///   needed the `[AGEND-MSG-PENDING]` hint. #1032 then migrated the
-///   two siblings — `emit_ci_conflict_alert` (#946, ~line 149) and
-///   the stale-SHA drop (#745, ~line 980). All three now use
-///   [`enqueue_with_idle_hint`]; the poller carries no remaining bare
-///   `enqueue` exceptions.
-/// - `mcp::handlers::instance` replace handover — the recipient is
-///   being killed and respawned, so any hint to the about-to-be-dead
-///   handle is a no-op and the fresh agent reads inbox on startup.
-///
-/// # Best-effort semantics
-///
-/// The PTY hint is fire-and-forget after the inbox write succeeds: a
-/// hint failure (recipient absent from registry, API unreachable) is
-/// swallowed silently because inbox storage is the durable source of
-/// truth.
-pub fn enqueue_with_idle_hint(home: &Path, target: &str, msg: InboxMessage) -> anyhow::Result<()> {
-    enqueue_with_idle_hint_with_emitter(home, target, msg, |hint| {
-        compose_aware_inject(home, target, hint);
-    })
-}
-
-/// Test-seam variant of [`enqueue_with_idle_hint`]. Accepts a closure
-/// that receives the formatted hint string so unit tests can verify
-/// the wire format without standing up the API loopback.
-pub(crate) fn enqueue_with_idle_hint_with_emitter<F>(
-    home: &Path,
-    target: &str,
-    mut msg: InboxMessage,
-    emit_hint: F,
-) -> anyhow::Result<()>
-where
-    F: FnOnce(&str),
-{
-    // Pre-stamp the id so the hint can reference it. enqueue() leaves an
-    // existing id untouched, so this stays a single auto-assignment.
-    ensure_msg_id(&mut msg);
-    let id = msg.id.clone().unwrap_or_default();
-    let from = msg.from.clone();
-    let kind = msg.kind.clone();
-
-    enqueue(home, target, msg)?;
-
-    // Build the PTY hint AFTER the inbox write succeeds — pending count
-    // reflects post-enqueue state so the recipient sees the same total
-    // they will read back via `inbox`.
-    let pending = unread_count(home, target).0;
-    let from_short = from.strip_prefix("from:").unwrap_or(&from);
-    let kind_str = kind.as_deref().unwrap_or("");
-    let hint = format!(
-        "{} id={} kind={} from={} inbox={} (use inbox tool)",
-        PENDING_HEADER_PREFIX,
-        sanitize_header_value(&id),
-        sanitize_header_value(kind_str),
-        sanitize_header_value(from_short),
-        pending,
-    );
-    emit_hint(&hint);
-    Ok(())
-}
-
-fn inject_with_submit(home: &Path, agent_name: &str, message: &str) -> anyhow::Result<()> {
-    let resp = crate::api::call(
-        home,
-        &serde_json::json!({
-            "method": crate::api::method::INJECT,
-            "params": {"name": agent_name, "data": message}
-        }),
-    )?;
-    if resp["ok"].as_bool() == Some(true) {
-        Ok(())
-    } else {
-        anyhow::bail!(
-            "{}",
-            resp["error"]
-                .as_str()
-                .unwrap_or("inject with submit failed")
-        );
-    }
-}
-
-/// #982 RC: submit-aware notification inject for the
-/// `notification_queue` flush path. Mirrors [`compose_aware_inject`]'s
-/// immediate-idle behaviour by routing through `INJECT` with no
-/// `raw: true` field — so the backend's `submit_key` is appended and
-/// the LLM model actually wakes up to consume the queued hint.
-///
-/// Background: the pre-fix flush path called a `raw=true` injector,
-/// which stranded `enqueue_with_idle_hint` hints that landed in the
-/// queue during a composing window. The hint text reached the
-/// recipient's prompt buffer but never got submitted, so codex
-/// one-shots (and any backend that needs explicit submit) silently
-/// dropped the wake.
-///
-/// Callers MUST be the queue-flush path (`app::flush_idle_notifications`)
-/// or an equivalent code path that has already gated on
-/// `notification_queue::is_composing` returning false — otherwise the
-/// submit-key inject would race with operator keystrokes.
-pub fn inject_notification_with_submit(
-    home: &Path,
-    agent_name: &str,
-    notification: &str,
-) -> anyhow::Result<()> {
-    inject_with_submit(home, agent_name, notification)
-}
-
-fn route_notification<F>(
-    home: &Path,
-    agent_name: &str,
-    notification: &str,
-    mut injector: F,
-) -> anyhow::Result<()>
-where
-    F: FnMut(&str) -> anyhow::Result<()>,
-{
-    if crate::notification_queue::is_composing(home, agent_name) {
-        crate::notification_queue::enqueue(home, agent_name, notification)?;
-        return Ok(());
-    }
-    injector(notification)
-}
-
-#[cfg(test)]
-#[allow(clippy::unwrap_used, clippy::expect_used)]
-mod tests {
-    use super::*;
-    use parking_lot::Mutex;
-    use std::fs;
-
-    /// Serializes tests that touch the global DISK_READONLY flag or rely on
-    /// enqueue not being blocked by it. Without this, `test_readonly_on_disk_full`
-    /// can set readonly=true and a concurrently-running enqueue test panics.
-    static READONLY_TEST_LOCK: Mutex<()> = Mutex::new(());
-
-    fn tmp_home(suffix: &str) -> PathBuf {
-        let dir =
-            std::env::temp_dir().join(format!("agend-inbox-{}-{}", suffix, std::process::id()));
-        fs::create_dir_all(&dir).ok();
-        dir
-    }
-
-    fn make_msg(from: &str, text: &str) -> InboxMessage {
-        InboxMessage {
-            from: from.to_string(),
-            text: text.to_string(),
-            timestamp: "2025-01-01T00:00:00Z".to_string(),
-            ..Default::default()
-        }
-    }
-
-    fn mark_composing(home: &Path, agent: &str) {
-        std::fs::create_dir_all(home.join("metadata")).ok();
-        std::fs::write(
-            home.join("metadata").join(format!("{agent}.json")),
-            format!(
-                "{{\"last_input_epoch_ms\":{}}}",
-                chrono::Utc::now().timestamp_millis()
-            ),
-        )
-        .ok();
-    }
-
-    #[test]
-    fn enqueue_drain_roundtrip() {
-        let home = tmp_home("roundtrip");
-        enqueue(&home, "agent1", make_msg("alice", "hello")).ok();
-        enqueue(&home, "agent1", make_msg("bob", "world")).ok();
-        enqueue(&home, "agent1", make_msg("carol", "!")).ok();
-
-        let msgs = drain(&home, "agent1");
-        assert_eq!(msgs.len(), 3);
-        assert_eq!(msgs[0].from, "alice");
-        assert_eq!(msgs[1].from, "bob");
-        assert_eq!(msgs[2].from, "carol");
-
-        fs::remove_dir_all(&home).ok();
-    }
-
-    #[test]
-    fn drain_empties_inbox() {
-        let home = tmp_home("drain-empty");
-        enqueue(&home, "agent1", make_msg("x", "y")).ok();
-
-        let first = drain(&home, "agent1");
-        assert_eq!(first.len(), 1);
-
-        let second = drain(&home, "agent1");
-        assert!(second.is_empty(), "second drain should be empty");
-
-        fs::remove_dir_all(&home).ok();
-    }
-
-    #[test]
-    fn drain_nonexistent_returns_empty() {
-        let home = tmp_home("no-inbox");
-        let msgs = drain(&home, "nonexistent");
-        assert!(msgs.is_empty());
-        fs::remove_dir_all(&home).ok();
-    }
-
-    #[test]
-    fn concurrent_enqueue_to_different_agents() {
-        let home = tmp_home("concurrent");
-        let home_arc = std::sync::Arc::new(home.clone());
-        let mut handles = vec![];
-
-        // Each thread writes to a different agent — no contention
-        for i in 0..10 {
-            let h = home_arc.clone();
-            handles.push(std::thread::spawn(move || {
-                let agent = format!("agent{i}");
-                enqueue(&h, &agent, make_msg(&format!("t{i}"), &format!("msg{i}")))
-                    .expect("enqueue should succeed");
-            }));
-        }
-        for h in handles {
-            h.join().expect("thread should not panic");
-        }
-
-        // Each agent should have exactly 1 message
-        for i in 0..10 {
-            let msgs = drain(&home, &format!("agent{i}"));
-            assert_eq!(msgs.len(), 1, "agent{i} should have 1 message");
-            assert_eq!(msgs[0].from, format!("t{i}"));
-        }
-
-        fs::remove_dir_all(&home).ok();
-    }
-
-    #[test]
-    fn notify_queues_when_composing() {
-        let home = tmp_home("notify-queue");
-        mark_composing(&home, "agent1");
-        let mut injected = false;
-        route_notification(&home, "agent1", "queued", |_| {
-            injected = true;
-            Ok(())
-        })
-        .expect("route should queue");
-        assert!(!injected);
-        assert_eq!(crate::notification_queue::pending_count(&home, "agent1"), 1);
-        fs::remove_dir_all(&home).ok();
-    }
-
-    #[test]
-    fn notify_injects_when_idle() {
-        let home = tmp_home("notify-idle");
-        let mut injected = Vec::new();
-        route_notification(&home, "agent1", "sent", |msg| {
-            injected.push(msg.to_string());
-            Ok(())
-        })
-        .expect("route should inject");
-        assert_eq!(injected, vec!["sent".to_string()]);
-        assert_eq!(crate::notification_queue::pending_count(&home, "agent1"), 0);
-        fs::remove_dir_all(&home).ok();
-    }
-
-    #[test]
-    fn inbox_message_fields_preserved() {
-        let home = tmp_home("fields");
-        let msg = InboxMessage {
-            schema_version: 0,
-            id: None,
-            read_at: None,
-            thread_id: None,
-            parent_id: None,
-            task_id: None,
-            force_meta: None,
-            correlation_id: None,
-            reviewed_head: None,
-            from: "sender".to_string(),
-            text: "body text".to_string(),
-            kind: Some("notification".to_string()),
-            timestamp: "2025-06-15T12:30:00Z".to_string(),
-            channel: None,
-            delivery_mode: None,
-            attachments: vec![],
-            in_reply_to_msg_id: None,
-            in_reply_to_excerpt: None,
-            superseded_by: None,
-            from_id: None,
-            broadcast_context: None,
-            sequencing: None,
-            eta_minutes: None,
-            reporting_cadence: None,
-            worktree_binding_required: None,
-            pr_number: None,
-        };
-        enqueue(&home, "agent1", msg).ok();
-        let msgs = drain(&home, "agent1");
-        assert_eq!(msgs.len(), 1);
-        assert_eq!(msgs[0].from, "sender");
-        assert_eq!(msgs[0].text, "body text");
-        assert_eq!(msgs[0].kind.as_deref(), Some("notification"));
-        assert_eq!(msgs[0].timestamp, "2025-06-15T12:30:00Z");
-
-        fs::remove_dir_all(&home).ok();
-    }
-
-    #[test]
-    fn deliver_short_message_also_enqueues() {
-        let home = tmp_home("deliver-short");
-        // deliver with short text — must enqueue to inbox for persistence
-        // (previously skipped enqueue for ≤500 chars, causing data loss)
-        deliver(
-            &home,
-            "agent1",
-            &NotifySource::Channel("user", crate::channel::ChannelKind::Telegram),
-            "short msg",
-            "\r",
-            None,
-            None,
-        );
-        let msgs = drain(&home, "agent1");
-        assert_eq!(
-            msgs.len(),
-            1,
-            "short messages must be enqueued for persistence"
-        );
-
-        fs::remove_dir_all(&home).ok();
-    }
-
-    #[test]
-    fn deliver_long_message_enqueues() {
-        let home = tmp_home("deliver-long");
-        let long_text: String = "x".repeat(600);
-        deliver(
-            &home,
-            "agent1",
-            &NotifySource::Channel("user", crate::channel::ChannelKind::Telegram),
-            &long_text,
-            "\r",
-            Some("chat".to_string()),
-            None,
-        );
-        let msgs = drain(&home, "agent1");
-        assert_eq!(msgs.len(), 1, "long messages should be enqueued");
-        assert_eq!(msgs[0].text, long_text);
-        assert_eq!(msgs[0].kind.as_deref(), Some("chat"));
-
-        fs::remove_dir_all(&home).ok();
-    }
-
-    #[test]
-    fn large_message_over_threshold() {
-        let home = tmp_home("large-msg");
-        let large_text: String = "a".repeat(10_000);
-        enqueue(&home, "agent1", make_msg("big", &large_text)).ok();
-        let msgs = drain(&home, "agent1");
-        assert_eq!(msgs.len(), 1);
-        assert_eq!(msgs[0].text.len(), 10_000);
-
-        fs::remove_dir_all(&home).ok();
-    }
-
-    #[test]
-    fn multiple_agents_isolated() {
-        let home = tmp_home("isolation");
-        enqueue(&home, "agent1", make_msg("a", "for-1")).ok();
-        enqueue(&home, "agent2", make_msg("b", "for-2")).ok();
-
-        let m1 = drain(&home, "agent1");
-        let m2 = drain(&home, "agent2");
-        assert_eq!(m1.len(), 1);
-        assert_eq!(m1[0].text, "for-1");
-        assert_eq!(m2.len(), 1);
-        assert_eq!(m2[0].text, "for-2");
-
-        fs::remove_dir_all(&home).ok();
-    }
-
-    #[test]
-    fn inbox_message_serialization() {
-        let msg = InboxMessage {
-            schema_version: 0,
-            id: None,
-            read_at: None,
-            thread_id: None,
-            parent_id: None,
-            task_id: None,
-            force_meta: None,
-            correlation_id: None,
-            reviewed_head: None,
-            from: "test".to_string(),
-            text: "hello \"world\"".to_string(),
-            kind: None,
-            timestamp: "2025-01-01T00:00:00Z".to_string(),
-            channel: None,
-            delivery_mode: None,
-            attachments: vec![],
-            in_reply_to_msg_id: None,
-            in_reply_to_excerpt: None,
-            superseded_by: None,
-            from_id: None,
-            broadcast_context: None,
-            sequencing: None,
-            eta_minutes: None,
-            reporting_cadence: None,
-            worktree_binding_required: None,
-            pr_number: None,
-        };
-        let json = serde_json::to_string(&msg).expect("serialize");
-        let parsed: InboxMessage = serde_json::from_str(&json).expect("deserialize");
-        assert_eq!(parsed.from, "test");
-        assert_eq!(parsed.text, "hello \"world\"");
-    }
-
-    #[test]
-    fn inbox_message_with_special_chars() {
-        let home = tmp_home("special");
-        let msg = InboxMessage {
-            schema_version: 0,
-            id: None,
-            read_at: None,
-            thread_id: None,
-            parent_id: None,
-            task_id: None,
-            force_meta: None,
-            correlation_id: None,
-            reviewed_head: None,
-            from: "user".to_string(),
-            text: "line1\nline2\ttab".to_string(),
-            kind: Some("special".to_string()),
-            timestamp: "2025-01-01T00:00:00Z".to_string(),
-            channel: None,
-            delivery_mode: None,
-            attachments: vec![],
-            in_reply_to_msg_id: None,
-            in_reply_to_excerpt: None,
-            superseded_by: None,
-            from_id: None,
-            broadcast_context: None,
-            sequencing: None,
-            eta_minutes: None,
-            reporting_cadence: None,
-            worktree_binding_required: None,
-            pr_number: None,
-        };
-        enqueue(&home, "agent1", msg).ok();
-        let msgs = drain(&home, "agent1");
-        assert_eq!(msgs.len(), 1);
-        assert_eq!(msgs[0].text, "line1\nline2\ttab");
-
-        fs::remove_dir_all(&home).ok();
-    }
-
-    // --- NotifySource tests ---
-
-    #[test]
-    fn notify_source_telegram_display() {
-        let s = NotifySource::Channel("chiacheng", crate::channel::ChannelKind::Telegram);
-        assert_eq!(s.to_string(), "user:chiacheng via telegram");
-        assert!(s.reply_hint().contains("reply tool"));
-    }
-
-    #[test]
-    fn notify_source_agent_display() {
-        let s = NotifySource::Agent("dev");
-        assert_eq!(s.to_string(), "from:dev");
-        let h = s.reply_hint();
-        assert!(h.contains("send tool"));
-        assert!(h.contains("dev"));
-    }
-
-    #[test]
-    fn notify_source_system_display() {
-        let s = NotifySource::System("ci");
-        assert_eq!(s.to_string(), "system:ci");
-        assert!(s.reply_hint().is_empty());
-    }
-
-    #[test]
-    fn drain_recovers_leftover_draining_file() {
-        // Simulates a crash between rename() and read: pending messages
-        // sit in `{name}.draining`. A second drain() must surface them,
-        // not drop them.
-        let home = tmp_home("recover");
-        let inbox_dir = home.join("inbox");
-        fs::create_dir_all(&inbox_dir).ok();
-        let draining = inbox_dir.join("agent1.draining");
-        let msg = serde_json::to_string(&make_msg("crashed", "pending")).expect("ser");
-        fs::write(&draining, format!("{msg}\n")).expect("write leftover");
-
-        let msgs = drain(&home, "agent1");
-        assert_eq!(msgs.len(), 1, "crashed batch must be recovered");
-        assert_eq!(msgs[0].from, "crashed");
-        assert_eq!(msgs[0].text, "pending");
-        // After successful read, leftover is cleared.
-        assert!(
-            !draining.exists(),
-            ".draining must be removed after successful drain"
-        );
-        fs::remove_dir_all(&home).ok();
-    }
-
-    #[test]
-    fn drain_does_not_overwrite_leftover_draining() {
-        // If a .draining file exists from a prior crash AND a new live
-        // inbox has arrived, the live file must be preserved — the new
-        // messages are picked up on the next drain cycle, not lost by a
-        // rename that overwrites the pending batch.
-        let home = tmp_home("no_overwrite");
-        let inbox_dir = home.join("inbox");
-        fs::create_dir_all(&inbox_dir).ok();
-
-        let draining = inbox_dir.join("agent1.draining");
-        let old_msg = serde_json::to_string(&make_msg("old", "from_crashed_batch")).expect("ser");
-        fs::write(&draining, format!("{old_msg}\n")).expect("write leftover");
-
-        enqueue(&home, "agent1", make_msg("new", "fresh")).ok();
-
-        // First drain: returns the crashed batch only.
-        let first = drain(&home, "agent1");
-        assert_eq!(first.len(), 1);
-        assert_eq!(first[0].from, "old");
-
-        // Second drain: picks up the new message now that .draining is gone.
-        let second = drain(&home, "agent1");
-        assert_eq!(second.len(), 1, "fresh message must survive recovery");
-        assert_eq!(second[0].from, "new");
-
-        fs::remove_dir_all(&home).ok();
-    }
-
-    #[test]
-    fn drain_read_failure_leaves_file_for_retry() {
-        // If read_to_string fails, .draining must remain on disk so a
-        // subsequent drain has another chance. (Simulating an unreadable
-        // file is awkward cross-platform; we instead assert the
-        // "retain-on-error" invariant by verifying successful drains
-        // DO remove, which is the inverse assertion our prior bug
-        // violated. See drain_recovers_leftover_draining_file.)
-        let home = tmp_home("retain");
-        let inbox_dir = home.join("inbox");
-        fs::create_dir_all(&inbox_dir).ok();
-        let draining = inbox_dir.join("agent1.draining");
-        // Non-UTF8 bytes → read_to_string returns Err.
-        fs::write(&draining, [0xFF, 0xFE, 0xFD]).expect("write");
-
-        let msgs = drain(&home, "agent1");
-        assert!(msgs.is_empty(), "unreadable batch yields no messages");
-        assert!(
-            draining.exists(),
-            ".draining must be retained after read failure for next retry"
-        );
-        fs::remove_dir_all(&home).ok();
-    }
-
-    #[test]
-    fn notify_agent_does_not_append_submit_key() {
-        // Verify the notification format doesn't contain \r (submit_key).
-        let source = NotifySource::Agent("peer");
-        let text = "hello world";
-        let display_text = text.to_string();
-        let notification = format!("[{source}] {display_text}{}", source.reply_hint());
-        assert!(
-            !notification.contains('\r'),
-            "notification must not contain submit_key (\\r): {notification:?}"
-        );
-    }
-
-    // -----------------------------------------------------------------------
-    // Regression pins: route_notification — the composing-aware primitive
-    // shared by `compose_aware_inject` and (pre-#1065) `compose_aware_send`.
-    // PR #96 conflated both into raw write; PR #99 split them. #1065 removed
-    // the now-redundant `compose_aware_send` wrapper, so these tests pin
-    // the underlying `route_notification` behavior directly.
-    // -----------------------------------------------------------------------
-
-    #[test]
-    fn route_notification_calls_injector_when_idle() {
-        // `route_notification` must call the injector closure (not enqueue
-        // to the notification_queue) when the agent is NOT composing.
-        let home = tmp_home("route-idle");
-        let mut called = false;
-        route_notification(&home, "agent1", "msg", |_| {
-            called = true;
-            Ok(())
-        })
-        .expect("route should call injector");
-        assert!(called, "injector must be called when agent is idle");
-        fs::remove_dir_all(&home).ok();
-    }
-
-    #[test]
-    fn route_notification_enqueues_when_composing() {
-        // `route_notification` must enqueue to the notification_queue (not
-        // call the injector) when the agent IS composing.
-        let home = tmp_home("route-composing");
-        mark_composing(&home, "agent1");
-        let mut called = false;
-        route_notification(&home, "agent1", "msg", |_| {
-            called = true;
-            Ok(())
-        })
-        .expect("route should enqueue");
-        assert!(!called, "injector must NOT be called when composing");
-        assert_eq!(
-            crate::notification_queue::pending_count(&home, "agent1"),
-            1,
-            "message must be enqueued"
-        );
-        fs::remove_dir_all(&home).ok();
-    }
-
-    #[test]
-    fn inject_with_submit_sends_raw_false() {
-        // Structural pin: inject_with_submit must NOT set raw=true in the
-        // INJECT API call. This ensures inject_to_agent (with submit_key)
-        // is used instead of write_to_agent (raw, no submit_key).
-        //
-        // We verify by inspecting the JSON payload construction. The function
-        // builds: {"method": "inject", "params": {"name": ..., "data": ...}}
-        // with NO "raw" field — handle_inject defaults raw=false → inject_to_agent.
-        //
-        // Cannot call inject_with_submit directly (needs running daemon), so
-        // we verify the contract structurally.
-        let send_json = serde_json::json!({
-            "method": crate::api::method::INJECT,
-            "params": {"name": "test", "data": "msg"}
-        });
-        assert!(
-            send_json["params"]["raw"].is_null(),
-            "inject_with_submit path must NOT set raw (defaults to false → inject_to_agent)"
-        );
-    }
-
-    #[test]
-    fn test_load_legacy_without_schema_version() {
-        let home = tmp_home("legacy-schema");
-        let inbox_dir = home.join("inbox");
-        fs::create_dir_all(&inbox_dir).ok();
-        // Write a legacy JSONL line without schema_version field
-        let legacy_line = r#"{"from":"old-agent","text":"legacy msg","kind":null,"timestamp":"2025-01-01T00:00:00Z"}"#;
-        fs::write(inbox_dir.join("agent1.jsonl"), format!("{legacy_line}\n")).ok();
-        let msgs = drain(&home, "agent1");
-        assert_eq!(msgs.len(), 1, "legacy message must load successfully");
-        assert_eq!(msgs[0].schema_version, 0, "missing field defaults to 0");
-        assert_eq!(msgs[0].from, "old-agent");
-        fs::remove_dir_all(&home).ok();
-    }
-
-    // --- Disk resilience tests ---
-
-    #[test]
-    fn test_readonly_on_disk_full() {
-        let _guard = READONLY_TEST_LOCK.lock();
-        // When DISK_READONLY is set, enqueue must fail and drain must still work.
-        let home = tmp_home("readonly");
-        enqueue(&home, "agent1", make_msg("a", "before")).ok();
-
-        DISK_READONLY.store(true, Ordering::Relaxed);
-        let result = enqueue(&home, "agent1", make_msg("b", "blocked"));
-        assert!(result.is_err(), "enqueue must fail in readonly mode");
-        assert!(
-            result.unwrap_err().to_string().contains("readonly"),
-            "error must mention readonly"
-        );
-
-        // drain still works in readonly mode
-        let msgs = drain(&home, "agent1");
-        assert_eq!(msgs.len(), 1);
-        assert_eq!(msgs[0].from, "a");
-
-        DISK_READONLY.store(false, Ordering::Relaxed);
-        fs::remove_dir_all(&home).ok();
-    }
-
-    #[test]
-    fn test_reject_future_schema_version() {
-        let home = tmp_home("future-schema");
-        let inbox_dir = home.join("inbox");
-        fs::create_dir_all(&inbox_dir).ok();
-        let future_line = r#"{"schema_version":999,"from":"future","text":"nope","kind":null,"timestamp":"2099-01-01T00:00:00Z"}"#;
-        let current_line = r#"{"schema_version":1,"from":"ok","text":"yes","kind":null,"timestamp":"2025-01-01T00:00:00Z"}"#;
-        fs::write(
-            inbox_dir.join("agent1.jsonl"),
-            format!("{future_line}\n{current_line}\n"),
-        )
-        .ok();
-        let msgs = drain(&home, "agent1");
-        assert_eq!(msgs.len(), 1, "future-versioned message must be rejected");
-        assert_eq!(msgs[0].from, "ok", "current-versioned message must survive");
-        fs::remove_dir_all(&home).ok();
-    }
-
-    #[test]
-    fn test_atomic_append_tmp_recovery() {
-        // Simulate a crash that left a .tmp file — recover_half_writes
-        // must move it to inbox.recovery/.
-        let home = tmp_home("atomic-recover");
-        let inbox_dir = home.join("inbox");
-        fs::create_dir_all(&inbox_dir).ok();
-
-        // Simulate stale tmp from interrupted enqueue
-        let tmp = inbox_dir.join("agent1.jsonl.tmp");
-        fs::write(
-            &tmp,
-            "{\"from\":\"x\",\"text\":\"orphan\",\"kind\":null,\"timestamp\":\"t\"}\n",
-        )
-        .ok();
-
-        recover_half_writes(&home);
-
-        assert!(!tmp.exists(), ".tmp must be moved to recovery");
-        let recovery = home.join("inbox.recovery");
-        assert!(recovery.exists(), "recovery dir must be created");
-        let entries: Vec<_> = fs::read_dir(&recovery).unwrap().flatten().collect();
-        assert_eq!(entries.len(), 1, "one timestamped recovery dir");
-
-        fs::remove_dir_all(&home).ok();
-    }
-
-    #[test]
-    fn test_half_written_jsonl_goes_to_recovery() {
-        // A JSONL file with a corrupt line must be moved to recovery.
-        let home = tmp_home("half-write");
-        let inbox_dir = home.join("inbox");
-        fs::create_dir_all(&inbox_dir).ok();
-
-        let jsonl = inbox_dir.join("agent1.jsonl");
-        let good = serde_json::to_string(&make_msg("ok", "fine")).unwrap();
-        // Write a good line followed by a truncated/corrupt line
-        fs::write(
-            &jsonl,
-            format!("{good}\n{{\"from\":\"broken\",\"text\":\"trun"),
-        )
-        .ok();
-
-        recover_half_writes(&home);
-
-        assert!(!jsonl.exists(), "corrupt JSONL must be moved to recovery");
-        let recovery = home.join("inbox.recovery");
-        assert!(recovery.exists());
-        // The recovery subdir should contain the moved file
-        let subdirs: Vec<_> = fs::read_dir(&recovery).unwrap().flatten().collect();
-        assert_eq!(subdirs.len(), 1);
-        let files: Vec<_> = fs::read_dir(subdirs[0].path()).unwrap().flatten().collect();
-        assert_eq!(files.len(), 1);
-        assert!(files[0].file_name().to_string_lossy().contains("agent1"));
-
-        fs::remove_dir_all(&home).ok();
-    }
-
-    #[test]
-    fn test_drain_marks_read_at_but_keeps_message() {
-        let home = tmp_home("drain-read-at");
-        enqueue(&home, "agent1", make_msg("alice", "hello")).ok();
-        enqueue(&home, "agent1", make_msg("bob", "world")).ok();
-
-        // First drain returns both messages with read_at set
-        let msgs = drain(&home, "agent1");
-        assert_eq!(msgs.len(), 2);
-        assert!(msgs[0].read_at.is_some(), "drain must stamp read_at");
-        assert!(msgs[1].read_at.is_some());
-
-        // Second drain returns empty (already read)
-        let msgs2 = drain(&home, "agent1");
-        assert!(
-            msgs2.is_empty(),
-            "already-read messages must not be returned"
-        );
-
-        // But the file still exists with the messages
-        let path = inbox_path(&home, "agent1");
-        let content = fs::read_to_string(&path).expect("file must still exist");
-        let lines: Vec<&str> = content.lines().filter(|l| !l.is_empty()).collect();
-        assert_eq!(lines.len(), 2, "messages must be kept in file");
-        // Verify read_at is persisted
-        let m: InboxMessage = serde_json::from_str(lines[0]).expect("parse");
-        assert!(m.read_at.is_some(), "read_at must be persisted to disk");
-
-        fs::remove_dir_all(&home).ok();
-    }
-
-    #[test]
-    fn test_sweep_expired_read_7d() {
-        let home = tmp_home("sweep-read-7d");
-        let inbox_dir = home.join("inbox");
-        fs::create_dir_all(&inbox_dir).ok();
-
-        let old_ts = (chrono::Utc::now() - chrono::Duration::days(10)).to_rfc3339();
-        let fresh_ts = chrono::Utc::now().to_rfc3339();
-        let read_old = format!(
-            r#"{{"schema_version":1,"id":"m-old","from":"a","text":"old read","kind":null,"timestamp":"{old_ts}","read_at":"{old_ts}"}}"#
-        );
-        let read_fresh = format!(
-            r#"{{"schema_version":1,"id":"m-fresh","from":"b","text":"fresh read","kind":null,"timestamp":"{fresh_ts}","read_at":"{fresh_ts}"}}"#
-        );
-        fs::write(
-            inbox_dir.join("agent1.jsonl"),
-            format!("{read_old}\n{read_fresh}\n"),
-        )
-        .ok();
-
-        sweep_expired(&home);
-
-        let content = fs::read_to_string(inbox_dir.join("agent1.jsonl")).expect("file");
-        let lines: Vec<&str> = content.lines().filter(|l| !l.is_empty()).collect();
-        assert_eq!(lines.len(), 1, "read message >7d must be swept");
-        assert!(
-            lines[0].contains("m-fresh"),
-            "fresh read message must survive"
-        );
-
-        fs::remove_dir_all(&home).ok();
-    }
-
-    #[test]
-    fn test_sweep_unread_30d() {
-        let home = tmp_home("sweep-unread-30d");
-        let inbox_dir = home.join("inbox");
-        fs::create_dir_all(&inbox_dir).ok();
-
-        let old_ts = (chrono::Utc::now() - chrono::Duration::days(35)).to_rfc3339();
-        let recent_ts = (chrono::Utc::now() - chrono::Duration::days(5)).to_rfc3339();
-        let unread_old = format!(
-            r#"{{"schema_version":1,"id":"m-unread-old","from":"a","text":"ancient","kind":null,"timestamp":"{old_ts}"}}"#
-        );
-        let unread_recent = format!(
-            r#"{{"schema_version":1,"id":"m-unread-recent","from":"b","text":"recent","kind":null,"timestamp":"{recent_ts}"}}"#
-        );
-        fs::write(
-            inbox_dir.join("agent1.jsonl"),
-            format!("{unread_old}\n{unread_recent}\n"),
-        )
-        .ok();
-
-        sweep_expired(&home);
-
-        let content = fs::read_to_string(inbox_dir.join("agent1.jsonl")).expect("file");
-        let lines: Vec<&str> = content.lines().filter(|l| !l.is_empty()).collect();
-        assert_eq!(lines.len(), 1, "unread message >30d must be swept");
-        assert!(
-            lines[0].contains("m-unread-recent"),
-            "recent unread must survive"
-        );
-
-        fs::remove_dir_all(&home).ok();
-    }
-
-    #[test]
-    fn test_describe_message_status_three_states() {
-        let home = tmp_home("describe-msg");
-        let inbox_dir = home.join("inbox");
-        fs::create_dir_all(&inbox_dir).ok();
-
-        let now = chrono::Utc::now().to_rfc3339();
-        let old_ts = (chrono::Utc::now() - chrono::Duration::days(35)).to_rfc3339();
-
-        // State 1: read message
-        let read_msg = format!(
-            r#"{{"schema_version":1,"id":"m-read","from":"a","text":"read","kind":null,"timestamp":"{now}","read_at":"{now}"}}"#
-        );
-        // State 2: unread expired (>30d)
-        let expired_msg = format!(
-            r#"{{"schema_version":1,"id":"m-expired","from":"b","text":"expired","kind":null,"timestamp":"{old_ts}"}}"#
-        );
-        fs::write(
-            inbox_dir.join("agent1.jsonl"),
-            format!("{read_msg}\n{expired_msg}\n"),
-        )
-        .ok();
-
-        // ReadAt
-        match describe_message(&home, "m-read", "agent1") {
-            MessageStatus::ReadAt(t, _dm) => assert_eq!(t, now),
-            other => panic!("expected ReadAt, got: {other:?}"),
-        }
-
-        // UnreadExpired
-        assert_eq!(
-            describe_message(&home, "m-expired", "agent1"),
-            MessageStatus::UnreadExpired
-        );
-
-        // NotFound
-        assert_eq!(
-            describe_message(&home, "m-nonexistent", "agent1"),
-            MessageStatus::NotFound
-        );
-
-        fs::remove_dir_all(&home).ok();
-    }
-
-    #[test]
-    fn test_enqueue_concurrent_same_agent() {
-        let _guard = READONLY_TEST_LOCK.lock();
-        let home = tmp_home("concurrent-same");
-        let home_arc = std::sync::Arc::new(home.clone());
-        let mut handles = vec![];
-
-        for i in 0..20 {
-            let h = home_arc.clone();
-            handles.push(std::thread::spawn(move || {
-                enqueue(&h, "agent1", make_msg(&format!("t{i}"), &format!("msg{i}")))
-                    .expect("enqueue should succeed");
-            }));
-        }
-        for h in handles {
-            h.join().expect("thread should not panic");
-        }
-
-        let msgs = drain(&home, "agent1");
-        assert_eq!(
-            msgs.len(),
-            20,
-            "all 20 concurrent enqueues must survive, got {}",
-            msgs.len()
-        );
-
-        fs::remove_dir_all(&home).ok();
-    }
-
-    #[test]
-    fn test_enqueue_vs_drain_no_lost_msg() {
-        let _guard = READONLY_TEST_LOCK.lock();
-        // Thread A enqueues 10 messages; thread B drains after each.
-        // Total drained must equal 10 — no lost messages.
-        let home = tmp_home("enqueue-vs-drain");
-        let home_a = std::sync::Arc::new(home.clone());
-        let home_b = home_a.clone();
-
-        let writer = std::thread::spawn(move || {
-            for i in 0..10 {
-                enqueue(
-                    &home_a,
-                    "agent1",
-                    make_msg(&format!("w{i}"), &format!("msg{i}")),
-                )
-                .expect("enqueue");
-                std::thread::sleep(std::time::Duration::from_millis(5));
-            }
-        });
-
-        let reader = std::thread::spawn(move || {
-            let mut total = Vec::new();
-            for _ in 0..20 {
-                let batch = drain(&home_b, "agent1");
-                total.extend(batch);
-                std::thread::sleep(std::time::Duration::from_millis(3));
-            }
-            total
-        });
-
-        writer.join().expect("writer");
-        let mut drained = reader.join().expect("reader");
-        // Final drain to catch any remaining
-        drained.extend(drain(&home, "agent1"));
-
-        assert_eq!(
-            drained.len(),
-            10,
-            "all 10 enqueued messages must be drained, got {}",
-            drained.len()
-        );
-
-        fs::remove_dir_all(&home).ok();
-    }
-
-    #[test]
-    fn test_concurrent_drain_no_duplicate_recovery() {
-        let _guard = READONLY_TEST_LOCK.lock();
-        // Pre-write a stale .draining file with 3 messages.
-        // Spawn 2 threads that both call drain simultaneously.
-        // Total recovered messages must be exactly 3 (no duplicates).
-        let home = tmp_home("concurrent-recovery");
-        let inbox_dir = home.join("inbox");
-        fs::create_dir_all(&inbox_dir).ok();
-
-        let draining = inbox_dir.join("agent1.draining");
-        let mut content = String::new();
-        for i in 0..3 {
-            let msg = make_msg(&format!("recover{i}"), &format!("msg{i}"));
-            content.push_str(&serde_json::to_string(&msg).unwrap());
-            content.push('\n');
-        }
-        fs::write(&draining, &content).ok();
-
-        let home_a = std::sync::Arc::new(home.clone());
-        let home_b = home_a.clone();
-
-        let a = std::thread::spawn(move || drain(&home_a, "agent1"));
-        let b = std::thread::spawn(move || drain(&home_b, "agent1"));
-
-        let mut all = a.join().expect("thread a");
-        all.extend(b.join().expect("thread b"));
-
-        assert_eq!(
-            all.len(),
-            3,
-            "exactly 3 recovered messages expected (no duplicates), got {}",
-            all.len()
-        );
-
-        fs::remove_dir_all(&home).ok();
-    }
-
-    #[test]
-    fn test_inbox_msg_thread_parent_fields_roundtrip() {
-        // New fields with #[serde(default)] must round-trip and be absent from legacy
-        let msg = InboxMessage {
-            schema_version: 1,
-            id: Some("m-1".into()),
-            from: "a".into(),
-            text: "t".into(),
-            kind: None,
-            timestamp: "2026-01-01T00:00:00Z".into(),
-            channel: None,
-            delivery_mode: None,
-            force_meta: None,
-            correlation_id: None,
-            reviewed_head: None,
-            read_at: None,
-            thread_id: Some("thread-42".into()),
-            parent_id: Some("m-0".into()),
-            task_id: None,
-            attachments: vec![],
-            in_reply_to_msg_id: None,
-            in_reply_to_excerpt: None,
-            superseded_by: None,
-            from_id: None,
-            broadcast_context: None,
-            sequencing: None,
-            eta_minutes: None,
-            reporting_cadence: None,
-            worktree_binding_required: None,
-            pr_number: None,
-        };
-        let json = serde_json::to_string(&msg).expect("ser");
-        assert!(json.contains("thread_id"));
-        assert!(json.contains("parent_id"));
-        let parsed: InboxMessage = serde_json::from_str(&json).expect("deser");
-        assert_eq!(parsed.thread_id.as_deref(), Some("thread-42"));
-        assert_eq!(parsed.parent_id.as_deref(), Some("m-0"));
-
-        // Legacy JSON without these fields → None (forward compat)
-        let legacy = r#"{"from":"x","text":"y","timestamp":"2026-01-01T00:00:00Z"}"#;
-        let parsed: InboxMessage = serde_json::from_str(legacy).expect("legacy deser");
-        assert!(parsed.thread_id.is_none());
-        assert!(parsed.parent_id.is_none());
-
-        // None fields should be omitted from serialization (skip_serializing_if)
-        let msg_no_thread = InboxMessage {
-            thread_id: None,
-            parent_id: None,
-            task_id: None,
-            force_meta: None,
-            correlation_id: None,
-            reviewed_head: None,
-            ..msg.clone()
-        };
-        let json2 = serde_json::to_string(&msg_no_thread).expect("ser");
-        assert!(!json2.contains("thread_id"));
-        assert!(!json2.contains("parent_id"));
-    }
-
-    #[test]
-    fn test_header_format_all_fields_present() {
-        let msg = InboxMessage {
-            schema_version: 1,
-            id: Some("m-42".into()),
-            from: "from:dev-lead".into(),
-            text: "x".repeat(500),
-            kind: Some("task".into()),
-            timestamp: "2026-01-01T00:00:00Z".into(),
-            channel: None,
-            delivery_mode: None,
-            force_meta: None,
-            correlation_id: None,
-            reviewed_head: None,
-            read_at: None,
-            thread_id: Some("t-100".into()),
-            parent_id: Some("m-41".into()),
-            task_id: None,
-            attachments: vec![],
-            in_reply_to_msg_id: None,
-            in_reply_to_excerpt: None,
-            superseded_by: None,
-            from_id: None,
-            broadcast_context: None,
-            sequencing: None,
-            eta_minutes: None,
-            reporting_cadence: None,
-            worktree_binding_required: None,
-            pr_number: None,
-        };
-        let header = format_header(&msg);
-        assert!(header.contains("[AGEND-MSG]"));
-        // #761: `from=` field strips the redundant `from:` prefix that
-        // the agent producer adds at Source::Agent display impl.
-        assert!(header.contains("from=dev-lead"));
-        assert!(!header.contains("from=from:"));
-        assert!(header.contains("id=m-42"));
-        assert!(header.contains("kind=task"));
-        assert!(header.contains("thread=t-100"));
-        assert!(header.contains("parent=m-41"));
-        assert!(header.contains("size=500"));
-        assert!(!header.contains('\n'), "header must be single line");
-    }
-
-    /// #761: agent-source `from:NAME` (sole producer at the
-    /// `Source::Agent` Display impl in inbox.rs:144) gets its redundant
-    /// `from:` prefix stripped at PTY-header rendering. Pre-fix the
-    /// header read `from=from:dev-fast-1`; the double prefix confused
-    /// agents that parsed the field as identity.
-    #[test]
-    fn test_header_format_strips_from_prefix_for_agent_sources() {
-        let msg = InboxMessage {
-            schema_version: 1,
-            id: Some("m-1".into()),
-            from: "from:dev-fast-1".into(),
-            text: "ping".into(),
-            kind: Some("task".into()),
-            timestamp: "2026-05-14T19:00:00Z".into(),
-            channel: None,
-            delivery_mode: None,
-            force_meta: None,
-            correlation_id: None,
-            reviewed_head: None,
-            read_at: None,
-            thread_id: None,
-            parent_id: None,
-            task_id: None,
-            attachments: vec![],
-            in_reply_to_msg_id: None,
-            in_reply_to_excerpt: None,
-            superseded_by: None,
-            from_id: None,
-            broadcast_context: None,
-            sequencing: None,
-            eta_minutes: None,
-            reporting_cadence: None,
-            worktree_binding_required: None,
-            pr_number: None,
-        };
-        let header = format_header(&msg);
-        assert!(
-            header.contains("from=dev-fast-1"),
-            "agent from= should drop the `from:` producer prefix: {header}"
-        );
-        assert!(
-            !header.contains("from=from:"),
-            "header must NOT carry the doubled `from=from:` form: {header}"
-        );
-    }
-
-    /// #761: non-agent sources (system events, telegram users) carry
-    /// their own namespace prefix (`system:`, `user:`) and must
-    /// survive the strip pass untouched. Only the literal `from:`
-    /// prefix from `Source::Agent` is dropped.
-    #[test]
-    fn test_header_format_preserves_system_namespace() {
-        let msg = InboxMessage {
-            schema_version: 1,
-            id: Some("m-1".into()),
-            from: "system:fleet_idle_watchdog".into(),
-            text: "ping".into(),
-            kind: Some("watchdog".into()),
-            timestamp: "2026-05-14T19:00:00Z".into(),
-            channel: None,
-            delivery_mode: None,
-            force_meta: None,
-            correlation_id: None,
-            reviewed_head: None,
-            read_at: None,
-            thread_id: None,
-            parent_id: None,
-            task_id: None,
-            attachments: vec![],
-            in_reply_to_msg_id: None,
-            in_reply_to_excerpt: None,
-            superseded_by: None,
-            from_id: None,
-            broadcast_context: None,
-            sequencing: None,
-            eta_minutes: None,
-            reporting_cadence: None,
-            worktree_binding_required: None,
-            pr_number: None,
-        };
-        let header = format_header(&msg);
-        assert!(
-            header.contains("from=system:fleet_idle_watchdog"),
-            "system namespace must survive untouched: {header}"
-        );
-    }
-
-    #[test]
-    fn test_header_format_omits_none_fields() {
-        let msg = InboxMessage {
-            schema_version: 1,
-            id: Some("m-1".into()),
-            from: "from:agent".into(),
-            text: "hello".into(),
-            kind: None,
-            timestamp: "2026-01-01T00:00:00Z".into(),
-            channel: None,
-            delivery_mode: None,
-            force_meta: None,
-            correlation_id: None,
-            reviewed_head: None,
-            read_at: None,
-            thread_id: None,
-            parent_id: None,
-            task_id: None,
-            attachments: vec![],
-            in_reply_to_msg_id: None,
-            in_reply_to_excerpt: None,
-            superseded_by: None,
-            from_id: None,
-            broadcast_context: None,
-            sequencing: None,
-            eta_minutes: None,
-            reporting_cadence: None,
-            worktree_binding_required: None,
-            pr_number: None,
-        };
-        let header = format_header(&msg);
-        // #761: agent-source prefix stripped.
-        assert!(header.contains("from=agent"));
-        assert!(!header.contains("from=from:"));
-        assert!(header.contains("id=m-1"));
-        assert!(!header.contains("kind="));
-        assert!(!header.contains("thread="));
-        assert!(!header.contains("parent="));
-        assert!(header.contains("size=5"));
-    }
-
-    #[test]
-    fn test_header_format_includes_attachments_when_present() {
-        let msg = InboxMessage {
-            schema_version: 1,
-            id: Some("m-1".into()),
-            from: "from:user".into(),
-            text: "see photo".into(),
-            kind: None,
-            timestamp: "2026-01-01T00:00:00Z".into(),
-            channel: None,
-            delivery_mode: None,
-            force_meta: None,
-            correlation_id: None,
-            reviewed_head: None,
-            read_at: None,
-            thread_id: None,
-            parent_id: None,
-            task_id: None,
-            attachments: vec![crate::channel::event::Attachment {
-                kind: crate::channel::event::AttachmentKind::Photo,
-                path: std::path::PathBuf::from("/tmp/photo.jpg"),
-                mime: None,
-                caption: None,
-                size_bytes: None,
-                original_filename: None,
-            }],
-            in_reply_to_msg_id: None,
-            in_reply_to_excerpt: None,
-            superseded_by: None,
-            from_id: None,
-            broadcast_context: None,
-            sequencing: None,
-            eta_minutes: None,
-            reporting_cadence: None,
-            worktree_binding_required: None,
-            pr_number: None,
-        };
-        let header = format_header(&msg);
-        assert!(
-            header.contains("attachments=[/tmp/photo.jpg]"),
-            "header must include attachment paths: {header}"
-        );
-    }
-
-    #[test]
-    fn test_header_format_omits_attachments_when_empty() {
-        let msg = InboxMessage {
-            schema_version: 1,
-            id: Some("m-1".into()),
-            from: "from:user".into(),
-            text: "text only".into(),
-            kind: None,
-            timestamp: "2026-01-01T00:00:00Z".into(),
-            channel: None,
-            delivery_mode: None,
-            force_meta: None,
-            correlation_id: None,
-            reviewed_head: None,
-            read_at: None,
-            thread_id: None,
-            parent_id: None,
-            task_id: None,
-            attachments: vec![],
-            in_reply_to_msg_id: None,
-            in_reply_to_excerpt: None,
-            superseded_by: None,
-            from_id: None,
-            broadcast_context: None,
-            sequencing: None,
-            eta_minutes: None,
-            reporting_cadence: None,
-            worktree_binding_required: None,
-            pr_number: None,
-        };
-        let header = format_header(&msg);
-        assert!(
-            !header.contains("attachments"),
-            "empty attachments must not appear in header: {header}"
-        );
-    }
-
-    #[test]
-    fn test_header_format_joins_multiple_attachments_with_comma() {
-        // Locks the `paths.join(",")` separator contract — future refactor that
-        // changes the separator (e.g. to ";" or " ") must update this test.
-        let mk_att = |p: &str| crate::channel::event::Attachment {
-            kind: crate::channel::event::AttachmentKind::Photo,
-            path: std::path::PathBuf::from(p),
-            mime: None,
-            caption: None,
-            size_bytes: None,
-            original_filename: None,
-        };
-        let msg = InboxMessage {
-            schema_version: 1,
-            id: Some("m-1".into()),
-            from: "from:user".into(),
-            text: "multi".into(),
-            kind: None,
-            timestamp: "2026-01-01T00:00:00Z".into(),
-            channel: None,
-            delivery_mode: None,
-            force_meta: None,
-            correlation_id: None,
-            reviewed_head: None,
-            read_at: None,
-            thread_id: None,
-            parent_id: None,
-            task_id: None,
-            attachments: vec![
-                mk_att("/tmp/a.jpg"),
-                mk_att("/tmp/b.jpg"),
-                mk_att("/tmp/c.jpg"),
-            ],
-            in_reply_to_msg_id: None,
-            in_reply_to_excerpt: None,
-            superseded_by: None,
-            from_id: None,
-            broadcast_context: None,
-            sequencing: None,
-            eta_minutes: None,
-            reporting_cadence: None,
-            worktree_binding_required: None,
-            pr_number: None,
-        };
-        let header = format_header(&msg);
-        assert!(
-            header.contains("attachments=[/tmp/a.jpg,/tmp/b.jpg,/tmp/c.jpg]"),
-            "multi-attachment paths must be comma-joined: {header}"
-        );
-    }
-
-    /// Sprint 54 layer-5: regression-proof for qa-test smoke
-    /// 2026-05-07 17:55 UTC — `send(team=qa-test, ...)` recipient PTY
-    /// header must surface the team= field so broadcast is
-    /// distinguishable from unicast at agent vantage. Mutate-revert:
-    /// removing the `format_header` broadcast-context branch leaves
-    /// this test failing with header lacking `team=` / `broadcast=`.
-    #[test]
-    fn test_header_format_includes_team_and_broadcast_when_team_broadcast() {
-        let msg = InboxMessage {
-            schema_version: 1,
-            id: Some("m-1".into()),
-            from: "from:lead".into(),
-            text: "ping".into(),
-            kind: Some("query".into()),
-            timestamp: "2026-05-07T18:00:00Z".into(),
-            channel: None,
-            delivery_mode: None,
-            force_meta: None,
-            correlation_id: None,
-            reviewed_head: None,
-            read_at: None,
-            thread_id: None,
-            parent_id: None,
-            task_id: None,
-            attachments: vec![],
-            in_reply_to_msg_id: None,
-            in_reply_to_excerpt: None,
-            superseded_by: None,
-            from_id: None,
-            broadcast_context: Some(BroadcastContext {
-                team: Some("qa-test".to_string()),
-                targets: vec!["kiro-cli-ea377a".into(), "kiro-cli-4e8a78".into()],
-                count: 2,
-            }),
-            sequencing: None,
-            eta_minutes: None,
-            reporting_cadence: None,
-            worktree_binding_required: None,
-            pr_number: None,
-        };
-        let header = format_header(&msg);
-        assert!(
-            header.contains("broadcast=2"),
-            "team broadcast must surface count: {header}"
-        );
-        assert!(
-            header.contains("team=qa-test"),
-            "team broadcast must surface team name (qa-test smoke success criteria): {header}"
-        );
-    }
-
-    /// Direct `targets=[…]` / `tags=[…]` fan-out has no team — header
-    /// gains `broadcast=N` only, no `team=`.
-    #[test]
-    fn test_header_format_includes_broadcast_only_when_targets_broadcast() {
-        let msg = InboxMessage {
-            schema_version: 1,
-            id: Some("m-1".into()),
-            from: "from:dev".into(),
-            text: "fyi".into(),
-            kind: None,
-            timestamp: "2026-05-07T18:00:00Z".into(),
-            channel: None,
-            delivery_mode: None,
-            force_meta: None,
-            correlation_id: None,
-            reviewed_head: None,
-            read_at: None,
-            thread_id: None,
-            parent_id: None,
-            task_id: None,
-            attachments: vec![],
-            in_reply_to_msg_id: None,
-            in_reply_to_excerpt: None,
-            superseded_by: None,
-            from_id: None,
-            broadcast_context: Some(BroadcastContext {
-                team: None,
-                targets: vec!["a".into(), "b".into(), "c".into()],
-                count: 3,
-            }),
-            sequencing: None,
-            eta_minutes: None,
-            reporting_cadence: None,
-            worktree_binding_required: None,
-            pr_number: None,
-        };
-        let header = format_header(&msg);
-        assert!(header.contains("broadcast=3"), "{header}");
-        assert!(
-            !header.contains("team="),
-            "no-team broadcast must not emit team= field: {header}"
-        );
-    }
-
-    /// Unicast must NOT surface broadcast/team fields — preserves the
-    /// pre-Sprint-54 header shape for non-broadcast callers (the
-    /// majority of SEND traffic).
-    #[test]
-    fn test_header_format_omits_broadcast_fields_when_unicast() {
-        let msg = InboxMessage {
-            schema_version: 1,
-            id: Some("m-1".into()),
-            from: "from:dev".into(),
-            text: "hi".into(),
-            kind: None,
-            timestamp: "2026-05-07T18:00:00Z".into(),
-            channel: None,
-            delivery_mode: None,
-            force_meta: None,
-            correlation_id: None,
-            reviewed_head: None,
-            read_at: None,
-            thread_id: None,
-            parent_id: None,
-            task_id: None,
-            attachments: vec![],
-            in_reply_to_msg_id: None,
-            in_reply_to_excerpt: None,
-            superseded_by: None,
-            from_id: None,
-            broadcast_context: None,
-            sequencing: None,
-            eta_minutes: None,
-            reporting_cadence: None,
-            worktree_binding_required: None,
-            pr_number: None,
-        };
-        let header = format_header(&msg);
-        assert!(
-            !header.contains("broadcast="),
-            "unicast must not emit broadcast= field: {header}"
-        );
-        assert!(
-            !header.contains("team="),
-            "unicast must not emit team= field: {header}"
-        );
-    }
-
-    /// Roundtrip: `broadcast_context` survives serde JSON encoding so the
-    /// inbox JSONL projection (visible via `inbox` MCP tool) matches the
-    /// PTY header. Absent field stays absent (no `null` leak).
-    #[test]
-    fn test_inbox_message_broadcast_context_serde_roundtrip() {
-        let with_ctx = InboxMessage {
-            schema_version: 1,
-            id: Some("m-1".into()),
-            from: "from:lead".into(),
-            text: "t".into(),
-            kind: None,
-            timestamp: "2026-05-07T18:00:00Z".into(),
-            channel: None,
-            delivery_mode: None,
-            force_meta: None,
-            correlation_id: None,
-            reviewed_head: None,
-            read_at: None,
-            thread_id: None,
-            parent_id: None,
-            task_id: None,
-            attachments: vec![],
-            in_reply_to_msg_id: None,
-            in_reply_to_excerpt: None,
-            superseded_by: None,
-            from_id: None,
-            broadcast_context: Some(BroadcastContext {
-                team: Some("qa-test".to_string()),
-                targets: vec!["a".into(), "b".into()],
-                count: 2,
-            }),
-            sequencing: None,
-            eta_minutes: None,
-            reporting_cadence: None,
-            worktree_binding_required: None,
-            pr_number: None,
-        };
-        let json = serde_json::to_string(&with_ctx).expect("ser");
-        assert!(json.contains("broadcast_context"));
-        assert!(json.contains("\"team\":\"qa-test\""));
-        let parsed: InboxMessage = serde_json::from_str(&json).expect("deser");
-        let parsed_ctx = parsed.broadcast_context.expect("ctx survived");
-        assert_eq!(parsed_ctx.team.as_deref(), Some("qa-test"));
-        assert_eq!(parsed_ctx.count, 2);
-        assert_eq!(parsed_ctx.targets, vec!["a", "b"]);
-
-        let without_ctx = InboxMessage {
-            broadcast_context: None,
-            sequencing: None,
-            eta_minutes: None,
-            reporting_cadence: None,
-            worktree_binding_required: None,
-            pr_number: None,
-            ..with_ctx
-        };
-        let json2 = serde_json::to_string(&without_ctx).expect("ser");
-        assert!(
-            !json2.contains("broadcast_context"),
-            "absent field must not serialize as null: {json2}"
-        );
-    }
-
-    #[test]
-    fn test_header_reply_excerpt_present() {
-        let mut msg = InboxMessage {
-            schema_version: 1,
-            id: Some("m-1".into()),
-            from: "u".into(),
-            text: "reply".into(),
-            kind: None,
-            timestamp: "t".into(),
-            channel: None,
-            delivery_mode: None,
-            force_meta: None,
-            correlation_id: None,
-            reviewed_head: None,
-            read_at: None,
-            thread_id: None,
-            parent_id: None,
-            task_id: None,
-            attachments: vec![],
-            in_reply_to_msg_id: Some("42".into()),
-            in_reply_to_excerpt: Some("[bob] original".into()),
-            superseded_by: None,
-            from_id: None,
-            broadcast_context: None,
-            sequencing: None,
-            eta_minutes: None,
-            reporting_cadence: None,
-            worktree_binding_required: None,
-            pr_number: None,
-        };
-        let h = format_header(&msg);
-        assert!(h.contains("reply_to_excerpt=[bob] original"), "{h}");
-        msg.in_reply_to_excerpt = None;
-        assert!(!format_header(&msg).contains("reply_to_excerpt"));
-    }
-
-    #[test]
-    fn test_reply_excerpt_long_truncated() {
-        let long: String = "x".repeat(300);
-        let trunc: String = long.chars().take(200).collect();
-        assert_eq!(trunc.len(), 200);
-        let excerpt = format!("[a] {trunc}…");
-        assert!(excerpt.contains("…"));
-    }
-
-    #[test]
-    fn test_build_excerpt_empty_returns_none() {
-        assert_eq!(build_excerpt("", "alice"), None);
-    }
-
-    #[test]
-    fn test_build_excerpt_short_text_no_ellipsis() {
-        let out = build_excerpt("hello", "alice").expect("non-empty text returns Some");
-        assert_eq!(out, "[alice] hello");
-        assert!(!out.contains('…'));
-    }
-
-    #[test]
-    fn test_build_excerpt_long_text_truncates_to_200_with_ellipsis() {
-        let long: String = "x".repeat(250);
-        let out = build_excerpt(&long, "bob").expect("non-empty text returns Some");
-        assert!(out.starts_with("[bob] "));
-        assert!(out.ends_with('…'));
-        // Body between "[bob] " prefix and trailing "…" must be 200 chars.
-        let body = out
-            .strip_prefix("[bob] ")
-            .and_then(|s| s.strip_suffix('…'))
-            .expect("expected `[bob] {200 chars}…` shape");
-        assert_eq!(body.chars().count(), 200);
-    }
-
-    #[test]
-    fn test_build_excerpt_cjk_truncated_to_200_chars() {
-        // 250 CJK chars (each is 3 bytes UTF-8). Verifies char-based
-        // truncation, not byte-based — a byte-based take(200) would slice
-        // mid-codepoint and produce invalid UTF-8.
-        let cjk: String = "一".repeat(250);
-        assert_eq!(cjk.chars().count(), 250);
-        assert_eq!(cjk.len(), 250 * 3);
-
-        let out = build_excerpt(&cjk, "carol").expect("non-empty text returns Some");
-        assert!(
-            std::str::from_utf8(out.as_bytes()).is_ok(),
-            "output must be valid UTF-8"
-        );
-        let body = out
-            .strip_prefix("[carol] ")
-            .and_then(|s| s.strip_suffix('…'))
-            .expect("expected `[carol] {200 CJK chars}…` shape");
-        assert_eq!(body.chars().count(), 200);
-        assert_eq!(body.len(), 200 * 3, "200 CJK chars = 600 UTF-8 bytes");
-    }
-
-    #[test]
-    fn test_build_excerpt_exactly_200_no_ellipsis() {
-        // Boundary: 200 chars exactly should NOT get truncated marker.
-        let exact: String = "y".repeat(200);
-        let out = build_excerpt(&exact, "dan").expect("non-empty text returns Some");
-        assert!(
-            !out.contains('…'),
-            "200-char input must not get ellipsis: {out}"
-        );
-    }
-
-    // Issue #672: inline-notification truncation hint must point at the
-    // `inbox` MCP tool, not at the (non-existent) `agend-terminal agent
-    // inbox` CLI subcommand. The CLI subcommand was never implemented;
-    // a user following the old hint hits `unrecognized subcommand`.
-    #[test]
-    fn test_inline_long_text_hint_points_at_mcp_tool_not_cli() {
-        let long: String = "x".repeat(250);
-        let out = format_notification_for_inject(false, &NotifySource::Agent("peer"), &long, &[]);
-        assert!(
-            out.contains("inbox MCP tool"),
-            "hint must direct caller to the MCP tool: {out}"
-        );
-        assert!(
-            !out.contains("agend-terminal agent"),
-            "hint must not reference the non-existent `agent` CLI subcommand: {out}"
-        );
-    }
-
-    #[test]
-    fn test_reply_excerpt_newline_escaped() {
-        let msg = InboxMessage {
-            schema_version: 1,
-            id: Some("m-1".into()),
-            from: "u".into(),
-            text: "r".into(),
-            kind: None,
-            timestamp: "t".into(),
-            channel: None,
-            delivery_mode: None,
-            force_meta: None,
-            correlation_id: None,
-            reviewed_head: None,
-            read_at: None,
-            thread_id: None,
-            parent_id: None,
-            task_id: None,
-            attachments: vec![],
-            in_reply_to_msg_id: Some("42".into()),
-            in_reply_to_excerpt: Some("[b] line1\nline2".into()),
-            superseded_by: None,
-            from_id: None,
-            broadcast_context: None,
-            sequencing: None,
-            eta_minutes: None,
-            reporting_cadence: None,
-            worktree_binding_required: None,
-            pr_number: None,
-        };
-        let h = format_header(&msg);
-        assert!(!h.contains('\n'), "must be single line: {h}");
-        assert!(h.contains("reply_to_excerpt="), "{h}");
-    }
-
-    #[test]
-    fn test_format_header_escapes_newlines() {
-        let msg = InboxMessage {
-            schema_version: 1,
-            id: Some("m-1".into()),
-            from: "from:evil\nagent".into(),
-            text: "hello".into(),
-            kind: Some("task\r\ninjection".into()),
-            timestamp: "2026-01-01T00:00:00Z".into(),
-            channel: None,
-            delivery_mode: None,
-            force_meta: None,
-            correlation_id: None,
-            reviewed_head: None,
-            read_at: None,
-            thread_id: Some("t\n1".into()),
-            parent_id: None,
-            task_id: None,
-            attachments: vec![],
-            in_reply_to_msg_id: None,
-            in_reply_to_excerpt: None,
-            superseded_by: None,
-            from_id: None,
-            broadcast_context: None,
-            sequencing: None,
-            eta_minutes: None,
-            reporting_cadence: None,
-            worktree_binding_required: None,
-            pr_number: None,
-        };
-        let header = format_header(&msg);
-        assert!(
-            !header.contains('\n'),
-            "header must not contain newline: {header:?}"
-        );
-        assert!(
-            !header.contains('\r'),
-            "header must not contain CR: {header:?}"
-        );
-        // #761: `from:` agent prefix is stripped BEFORE sanitize, so the
-        // newline-escaped tail `evil agent` is what surfaces in the field.
-        assert!(header.contains("from=evil agent"));
-        assert!(!header.contains("from=from:"));
-        assert!(header.contains("kind=task  injection"));
-    }
-
-    #[test]
-    fn test_format_header_escapes_control_chars() {
-        let msg = InboxMessage {
-            schema_version: 1,
-            id: None,
-            from: "from:\x00null\x07bell".into(),
-            text: "t".into(),
-            kind: None,
-            timestamp: "2026-01-01T00:00:00Z".into(),
-            channel: None,
-            delivery_mode: None,
-            force_meta: None,
-            correlation_id: None,
-            reviewed_head: None,
-            read_at: None,
-            thread_id: None,
-            parent_id: None,
-            task_id: None,
-            attachments: vec![],
-            in_reply_to_msg_id: None,
-            in_reply_to_excerpt: None,
-            superseded_by: None,
-            from_id: None,
-            broadcast_context: None,
-            sequencing: None,
-            eta_minutes: None,
-            reporting_cadence: None,
-            worktree_binding_required: None,
-            pr_number: None,
-        };
-        let header = format_header(&msg);
-        assert!(
-            !header.chars().any(|c| c.is_control() && c != '\x1b'),
-            "header must not contain control chars (except ANSI escape): {header:?}"
-        );
-    }
-
-    #[test]
-    fn test_short_msg_below_threshold() {
-        // Messages <= HEADER_SIZE_THRESHOLD should NOT use header format
-        let short = "a".repeat(HEADER_SIZE_THRESHOLD);
-        assert!(short.len() <= HEADER_SIZE_THRESHOLD);
-    }
-
-    #[test]
-    fn test_long_msg_above_threshold() {
-        // Messages > HEADER_SIZE_THRESHOLD should use header-only injection
-        let long = "a".repeat(HEADER_SIZE_THRESHOLD + 1);
-        assert!(long.len() > HEADER_SIZE_THRESHOLD);
-        // format_header produces a compact single-line representation
-        let msg = InboxMessage {
-            schema_version: 1,
-            id: Some("m-1".into()),
-            from: "from:x".into(),
-            text: long,
-            kind: Some("task".into()),
-            timestamp: "2026-01-01T00:00:00Z".into(),
-            channel: None,
-            delivery_mode: None,
-            force_meta: None,
-            correlation_id: None,
-            reviewed_head: None,
-            read_at: None,
-            thread_id: None,
-            parent_id: None,
-            task_id: None,
-            attachments: vec![],
-            in_reply_to_msg_id: None,
-            in_reply_to_excerpt: None,
-            superseded_by: None,
-            from_id: None,
-            broadcast_context: None,
-            sequencing: None,
-            eta_minutes: None,
-            reporting_cadence: None,
-            worktree_binding_required: None,
-            pr_number: None,
-        };
-        let header = format_header(&msg);
-        assert!(
-            header.len() < HEADER_SIZE_THRESHOLD,
-            "header must be compact"
-        );
-        assert!(header.contains(&format!("size={}", HEADER_SIZE_THRESHOLD + 1)));
-    }
-
-    #[test]
-    fn test_threshold_uses_char_count_not_bytes() {
-        // 100 CJK chars = 100 chars but 300 bytes (3 bytes each in UTF-8).
-        // Must be treated as short (100 < 300 threshold), not long.
-        let cjk = "你".repeat(100);
-        assert_eq!(cjk.chars().count(), 100);
-        assert_eq!(cjk.len(), 300); // bytes
-                                    // 100 chars < HEADER_SIZE_THRESHOLD (300) → should be short path
-        assert!(cjk.chars().count() <= HEADER_SIZE_THRESHOLD);
-
-        // 301 CJK chars = 301 chars → should be long path
-        let long_cjk = "你".repeat(HEADER_SIZE_THRESHOLD + 1);
-        assert!(long_cjk.chars().count() > HEADER_SIZE_THRESHOLD);
-
-        // format_header size= should report char count, not byte count
-        let msg = InboxMessage {
-            schema_version: 1,
-            id: None,
-            from: "from:x".into(),
-            text: cjk,
-            kind: None,
-            timestamp: "2026-01-01T00:00:00Z".into(),
-            channel: None,
-            delivery_mode: None,
-            force_meta: None,
-            correlation_id: None,
-            reviewed_head: None,
-            read_at: None,
-            thread_id: None,
-            parent_id: None,
-            task_id: None,
-            attachments: vec![],
-            in_reply_to_msg_id: None,
-            in_reply_to_excerpt: None,
-            superseded_by: None,
-            from_id: None,
-            broadcast_context: None,
-            sequencing: None,
-            eta_minutes: None,
-            reporting_cadence: None,
-            worktree_binding_required: None,
-            pr_number: None,
-        };
-        let header = format_header(&msg);
-        assert!(
-            header.contains("size=100"),
-            "size must be char count (100), not byte count (300): {header}"
-        );
-    }
-
-    #[test]
-    fn test_format_event_header_basic() {
-        let h = format_event_header("poll-reminder", &[("unread", "3"), ("oldest", "5m")]);
-        assert!(h.contains("[AGEND-MSG]"), "must have prefix");
-        assert!(h.contains("kind=poll-reminder"), "must have kind");
-        assert!(h.contains("unread=3"), "must have unread field");
-        assert!(h.contains("oldest=5m"), "must have oldest field");
-        assert!(!h.contains('\n'), "must be single line");
-    }
-
-    #[test]
-    fn test_format_event_header_sanitizes_fields() {
-        let h = format_event_header("evil\nkind", &[("key", "val\r\nue")]);
-        assert!(
-            !h.contains('\n'),
-            "newlines in kind must be sanitized: {h:?}"
-        );
-        assert!(!h.contains('\r'), "CR in value must be sanitized: {h:?}");
-        assert!(h.contains("kind=evil kind"));
-        assert!(h.contains("key=val  ue"));
-    }
-
-    /// Serialize tests that mutate AGEND_POINTER_ONLY_INJECT env var.
-    static ENV_LOCK: parking_lot::Mutex<()> = parking_lot::Mutex::new(());
-
-    #[test]
-    fn test_pointer_only_feature_flag() {
-        let _lock = ENV_LOCK.lock();
-        std::env::set_var("AGEND_POINTER_ONLY_INJECT", "1");
-        assert!(pointer_only_inject(), "flag=1 must enable pointer-only");
-        std::env::remove_var("AGEND_POINTER_ONLY_INJECT");
-    }
-
-    #[test]
-    fn test_pointer_only_disabled_default() {
-        let _lock = ENV_LOCK.lock();
-        std::env::remove_var("AGEND_POINTER_ONLY_INJECT");
-        assert!(!pointer_only_inject(), "unset flag must default to false");
-        std::env::set_var("AGEND_POINTER_ONLY_INJECT", "0");
-        assert!(!pointer_only_inject(), "flag=0 must be disabled");
-        std::env::remove_var("AGEND_POINTER_ONLY_INJECT");
-    }
-
-    #[test]
-    fn test_notify_agent_pointer_only_emits_header_not_body() {
-        // Pure function test: pointer_only=true → output contains [AGEND-MSG]
-        // + size= but NOT the body text.
-        let source = NotifySource::Agent("sender");
-        let s = format_notification_for_inject(true, &source, "secret body text", &[]);
-        assert!(
-            s.contains("[AGEND-MSG]"),
-            "pointer mode must contain [AGEND-MSG]: {s}"
-        );
-        assert!(s.contains("size="), "pointer mode must contain size=: {s}");
-        assert!(
-            !s.contains("secret body text"),
-            "pointer mode must NOT contain body: {s}"
-        );
-        assert!(
-            s.contains("use inbox tool"),
-            "pointer mode must direct to inbox: {s}"
-        );
-    }
-
-    #[test]
-    fn test_notify_agent_default_inline_behavior() {
-        // Pure function test: pointer_only=false → output contains the body text.
-        let source = NotifySource::Agent("sender");
-        let s = format_notification_for_inject(false, &source, "inline text", &[]);
-        assert!(
-            s.contains("inline text"),
-            "default mode must contain body: {s}"
-        );
-        assert!(
-            !s.contains("[AGEND-MSG]"),
-            "default mode must NOT contain [AGEND-MSG] header: {s}"
-        );
-    }
-
-    #[test]
-    fn inbox_message_typed_channel_field_compat() {
-        // New typed channel field roundtrip
-        let mut msg = make_msg("test", "hello");
-        assert!(msg.channel.is_none());
-        msg.channel = Some(crate::channel::ChannelKind::Telegram);
-        let serialized = serde_json::to_string(&msg).unwrap();
-        assert!(
-            serialized.contains(r#""channel":"telegram""#),
-            "channel must serialize as snake_case: {serialized}"
-        );
-        let reparsed: InboxMessage = serde_json::from_str(&serialized).unwrap();
-        assert_eq!(
-            reparsed.channel,
-            Some(crate::channel::ChannelKind::Telegram)
-        );
-
-        // Legacy JSONL without channel field → None
-        let legacy =
-            r#"{"schema_version":1,"from":"x","text":"y","timestamp":"2026-01-01T00:00:00Z"}"#;
-        let legacy_msg: InboxMessage = serde_json::from_str(legacy).unwrap();
-        assert_eq!(legacy_msg.channel, None);
-    }
-
-    #[test]
-    fn legacy_inbox_message_without_attachments_deserializes() {
-        // Old messages (pre-PR-AF) have no `attachments` field.
-        // serde(default) must fill Vec::new() so deserialization succeeds.
-        let legacy = r#"{"from":"user:op","text":"hello","timestamp":"2026-04-26T00:00:00Z"}"#;
-        let msg: InboxMessage = serde_json::from_str(legacy).unwrap();
-        assert!(msg.attachments.is_empty());
-    }
-
-    #[test]
-    fn legacy_inbox_message_without_excerpt_deserializes() {
-        // Old messages (pre-PR-AQ) have no `in_reply_to_excerpt` field.
-        // serde(default) must fill `None` so deserialization succeeds — locks
-        // schema backward-compatibility for on-disk JSONL written before PR-AQ.
-        let legacy = r#"{"from":"user:op","text":"hello","timestamp":"2026-04-26T00:00:00Z"}"#;
-        let msg: InboxMessage = serde_json::from_str(legacy).unwrap();
-        assert_eq!(msg.in_reply_to_excerpt, None);
-        assert_eq!(msg.in_reply_to_msg_id, None);
-    }
-
-    #[test]
-    fn inbox_message_with_attachment_roundtrips() {
-        use crate::channel::event::{Attachment, AttachmentKind};
-        let msg = InboxMessage {
-            schema_version: 1,
-            id: None,
-            read_at: None,
-            thread_id: None,
-            parent_id: None,
-            task_id: None,
-            force_meta: None,
-            correlation_id: None,
-            reviewed_head: None,
-            from: "user:op".into(),
-            text: "see photo".into(),
-            kind: None,
-            timestamp: "2026-04-26T00:00:00Z".into(),
-            channel: None,
-            delivery_mode: None,
-            attachments: vec![Attachment {
-                kind: AttachmentKind::Photo,
-                path: "/tmp/photo.jpg".into(),
-                mime: Some("image/jpeg".into()),
-                caption: Some("test".into()),
-                size_bytes: Some(1234),
-                original_filename: None,
-            }],
-            in_reply_to_msg_id: None,
-            in_reply_to_excerpt: None,
-            superseded_by: None,
-            from_id: None,
-            broadcast_context: None,
-            sequencing: None,
-            eta_minutes: None,
-            reporting_cadence: None,
-            worktree_binding_required: None,
-            pr_number: None,
-        };
-        let json = serde_json::to_string(&msg).unwrap();
-        let back: InboxMessage = serde_json::from_str(&json).unwrap();
-        assert_eq!(back.attachments.len(), 1);
-        assert_eq!(back.attachments[0].kind, AttachmentKind::Photo);
-        assert_eq!(back.attachments[0].size_bytes, Some(1234));
-    }
-
-    #[test]
-    fn inbox_message_with_in_reply_to_msg_id_roundtrips() {
-        let mut msg = make_msg("user:op", "reply test");
-        msg.in_reply_to_msg_id = Some("999".to_string());
-        let json = serde_json::to_string(&msg).unwrap();
-        assert!(
-            json.contains(r#""in_reply_to_msg_id":"999""#),
-            "json: {json}"
-        );
-        let back: InboxMessage = serde_json::from_str(&json).unwrap();
-        assert_eq!(back.in_reply_to_msg_id, Some("999".to_string()));
-    }
-
-    // --- M6: ci-watch supersede tests ---
-
-    #[test]
-    fn superseded_by_field_backward_compat() {
-        // Old JSONL without superseded_by should deserialize with None
-        let json = r#"{"from":"system:ci","text":"test","timestamp":"2026-01-01T00:00:00Z"}"#;
-        let msg: InboxMessage = serde_json::from_str(json).unwrap();
-        assert!(msg.superseded_by.is_none());
-    }
-
-    #[test]
-    fn mark_ci_watch_superseded_tags_prior_messages() {
-        let home = tmp_home("supersede_tag");
-        let agent = "test-agent";
-        let msg1 = InboxMessage {
-            schema_version: 0,
-            id: Some("old-1".into()),
-            from: "system:ci".into(),
-            text: "[ci-pass] owner/repo@main: passed ✓".into(),
-            kind: Some("ci-watch".into()),
-            timestamp: "2026-01-01T00:00:00Z".into(),
-            read_at: None,
-            thread_id: None,
-            parent_id: None,
-            task_id: None,
-            force_meta: None,
-            correlation_id: None,
-            reviewed_head: None,
-            channel: None,
-            delivery_mode: None,
-            attachments: vec![],
-            in_reply_to_msg_id: None,
-            in_reply_to_excerpt: None,
-            superseded_by: None,
-            from_id: None,
-            broadcast_context: None,
-            sequencing: None,
-            eta_minutes: None,
-            reporting_cadence: None,
-            worktree_binding_required: None,
-            pr_number: None,
-        };
-        enqueue(&home, agent, msg1).unwrap();
-        mark_ci_watch_superseded(&home, agent, "owner/repo@main", "new-msg-id");
-        let msgs = drain(&home, agent);
-        assert!(
-            msgs.is_empty(),
-            "superseded message should be filtered: {msgs:?}"
-        );
-        std::fs::remove_dir_all(&home).ok();
-    }
-
-    #[test]
-    fn drain_excludes_superseded_messages() {
-        let home = tmp_home("drain_superseded");
-        let agent = "test-agent";
-        let normal = InboxMessage {
-            schema_version: 0,
-            id: Some("normal-1".into()),
-            from: "from:lead".into(),
-            text: "hello".into(),
-            kind: None,
-            timestamp: "2026-01-01T00:00:00Z".into(),
-            read_at: None,
-            thread_id: None,
-            parent_id: None,
-            task_id: None,
-            force_meta: None,
-            correlation_id: None,
-            reviewed_head: None,
-            channel: None,
-            delivery_mode: None,
-            attachments: vec![],
-            in_reply_to_msg_id: None,
-            in_reply_to_excerpt: None,
-            superseded_by: None,
-            from_id: None,
-            broadcast_context: None,
-            sequencing: None,
-            eta_minutes: None,
-            reporting_cadence: None,
-            worktree_binding_required: None,
-            pr_number: None,
-        };
-        let superseded = InboxMessage {
-            schema_version: 0,
-            id: Some("old-ci".into()),
-            from: "system:ci".into(),
-            text: "[ci-pass] repo@main".into(),
-            kind: Some("ci-watch".into()),
-            timestamp: "2026-01-01T00:00:01Z".into(),
-            read_at: None,
-            thread_id: None,
-            parent_id: None,
-            task_id: None,
-            force_meta: None,
-            correlation_id: None,
-            reviewed_head: None,
-            channel: None,
-            delivery_mode: None,
-            attachments: vec![],
-            in_reply_to_msg_id: None,
-            in_reply_to_excerpt: None,
-            superseded_by: Some("new-ci".into()),
-            from_id: None,
-            broadcast_context: None,
-            sequencing: None,
-            eta_minutes: None,
-            reporting_cadence: None,
-            worktree_binding_required: None,
-            pr_number: None,
-        };
-        enqueue(&home, agent, normal).unwrap();
-        enqueue(&home, agent, superseded).unwrap();
-        let msgs = drain(&home, agent);
-        assert_eq!(msgs.len(), 1, "only non-superseded: {msgs:?}");
-        assert_eq!(msgs[0].id.as_deref(), Some("normal-1"));
-        std::fs::remove_dir_all(&home).ok();
-    }
-
-    #[test]
-    fn from_id_round_trip() {
-        let msg = InboxMessage {
-            schema_version: 0,
-            id: Some("m-1".into()),
-            from: "from:dev".into(),
-            from_id: Some("a3k9p2xf".into()),
-            text: "hello".into(),
-            kind: None,
-            timestamp: "2026-01-01T00:00:00Z".into(),
-            read_at: None,
-            thread_id: None,
-            parent_id: None,
-            task_id: None,
-            force_meta: None,
-            correlation_id: None,
-            reviewed_head: None,
-            channel: None,
-            delivery_mode: None,
-            attachments: vec![],
-            in_reply_to_msg_id: None,
-            in_reply_to_excerpt: None,
-            superseded_by: None,
-            broadcast_context: None,
-            sequencing: None,
-            eta_minutes: None,
-            reporting_cadence: None,
-            worktree_binding_required: None,
-            pr_number: None,
-        };
-        let json = serde_json::to_string(&msg).expect("serialize");
-        let parsed: InboxMessage = serde_json::from_str(&json).expect("deserialize");
-        assert_eq!(parsed.from_id, Some("a3k9p2xf".to_string()));
-    }
-
-    // ── Sprint 54 silent-drop layer-4 hotfix: PTY attachment indicator ──
-    //
-    // Operator m-9 dispatch m-20260507131404761875-8. The pre-hotfix
-    // `format_notification_for_inject` ignored attachments entirely:
-    //   - pointer_only=true header had no `attachments=[…]` field
-    //   - pointer_only=false body became empty when text was empty,
-    //     even with attachments present (silent-drop class 4th instance,
-    //     decision `d-20260507125347359886-0`)
-    //
-    // Each test pins one of the contract gates from the dispatch.
-    //
-    // EMPIRICAL REGRESSION-PROOF ANCHOR: replacing
-    // `summarize_attachments_for_header` to always return `None` AND
-    // collapsing the placeholder branch in `format_notification_for_inject`
-    // makes both `…header_carries_attachments_field…` and
-    // `…inline_body_substitutes_placeholder_when_text_empty…` fail.
-    // PR description carries the verbatim FAIL signatures.
-
-    fn p54_attachment(
-        kind: crate::channel::event::AttachmentKind,
-    ) -> crate::channel::event::Attachment {
-        crate::channel::event::Attachment {
-            kind,
-            path: std::path::PathBuf::from("/tmp/p54-fixture"),
-            mime: None,
-            caption: None,
-            size_bytes: Some(50_000),
-            original_filename: None,
-        }
-    }
-
-    fn p54_attachment_with_name(
-        kind: crate::channel::event::AttachmentKind,
-        name: &str,
-    ) -> crate::channel::event::Attachment {
-        let mut a = p54_attachment(kind);
-        a.original_filename = Some(name.to_string());
-        a
-    }
-
-    #[test]
-    fn summarize_attachments_for_header_aggregates_by_kind() {
-        // Layer-4 gate: header `attachments=[…]` field is kind-aggregated
-        // counts in stable order — operator m-9 spec literal:
-        // `attachments=[1 photo, 2 document]`.
-        use crate::channel::event::AttachmentKind;
-        let attachments = vec![
-            p54_attachment(AttachmentKind::Photo),
-            p54_attachment(AttachmentKind::Document),
-            p54_attachment(AttachmentKind::Document),
-        ];
-        let summary = summarize_attachments_for_header(&attachments)
-            .expect("non-empty input must return Some");
-        assert_eq!(
-            summary, "1 photo, 2 document",
-            "kind-aggregated stable order"
-        );
-    }
-
-    #[test]
-    fn summarize_attachments_for_header_returns_none_for_empty() {
-        // Edge: empty input → None so callers don't emit
-        // `attachments=[]` for non-attachment notifications.
-        assert!(summarize_attachments_for_header(&[]).is_none());
-    }
-
-    #[test]
-    fn pointer_only_header_carries_attachments_field_when_present() {
-        // Layer-4 gate (regression-proof anchor): pointer_only=true
-        // header gains `attachments=[…]` when attachments are present.
-        // Pre-r0 produced `[AGEND-MSG] size=N (use inbox tool)` only —
-        // agent had no signal that media was waiting.
-        use crate::channel::event::AttachmentKind;
-        let attachments = vec![
-            p54_attachment(AttachmentKind::Photo),
-            p54_attachment(AttachmentKind::Document),
-        ];
-        let source = NotifySource::Channel("alice", crate::channel::ChannelKind::Telegram);
-        let s = format_notification_for_inject(true, &source, "", &attachments);
-        assert!(
-            s.contains("[AGEND-MSG]"),
-            "pointer mode must keep [AGEND-MSG]: {s}"
-        );
-        assert!(
-            s.contains("attachments=[1 photo, 1 document]"),
-            "header must carry attachments summary: {s}"
-        );
-    }
-
-    #[test]
-    fn pointer_only_header_omits_attachments_field_when_empty() {
-        // Layer-4 gate: empty attachments → no field. Avoids polluting
-        // existing notifications (operator alerts, status keywords)
-        // with an empty marker.
-        let source = NotifySource::Agent("sender");
-        let s = format_notification_for_inject(true, &source, "hello", &[]);
-        assert!(
-            !s.contains("attachments=["),
-            "no attachments field for empty input: {s}"
-        );
-    }
-
-    #[test]
-    fn inline_body_substitutes_placeholder_when_text_empty_with_attachments() {
-        // Layer-4 gate (regression-proof anchor): pointer_only=false +
-        // text="" + attachments non-empty produces a human-readable
-        // placeholder instead of a content-less inline notification.
-        // Without this, the agent received `[user:foo via telegram] `
-        // — empty after the source tag — and had nothing to action on.
-        use crate::channel::event::AttachmentKind;
-        let attachments = vec![p54_attachment_with_name(AttachmentKind::Photo, "cat.jpg")];
-        let source = NotifySource::Channel("alice", crate::channel::ChannelKind::Telegram);
-        let s = format_notification_for_inject(false, &source, "", &attachments);
-        assert!(
-            s.contains("[1 photo: cat.jpg]"),
-            "single-attachment placeholder uses filename: {s}"
-        );
-    }
-
-    #[test]
-    fn inline_body_keeps_text_when_present_with_attachments() {
-        // Layer-4 gate: caption present → text passes through. The
-        // placeholder must NEVER overwrite the user's own words.
-        // Mirrors layer-2 #497 invariant.
-        use crate::channel::event::AttachmentKind;
-        let attachments = vec![p54_attachment(AttachmentKind::Photo)];
-        let source = NotifySource::Channel("alice", crate::channel::ChannelKind::Telegram);
-        let s = format_notification_for_inject(false, &source, "look at this!", &attachments);
-        assert!(
-            s.contains("look at this!"),
-            "caption must pass through: {s}"
-        );
-        assert!(
-            !s.contains("[1 photo"),
-            "placeholder must NOT overwrite caption: {s}"
-        );
-    }
-
-    #[test]
-    fn attachment_body_placeholder_multi_uses_aggregated_summary() {
-        // Layer-4 gate: multi-attachment placeholder uses kind summary
-        // rather than per-file enumeration. Single attachment with no
-        // filename also uses the summary form ("[1 photo attached]").
-        use crate::channel::event::AttachmentKind;
-        let multi = vec![
-            p54_attachment(AttachmentKind::Photo),
-            p54_attachment(AttachmentKind::Document),
-            p54_attachment(AttachmentKind::Document),
-        ];
-        let s = attachment_body_placeholder(&multi);
-        assert_eq!(s, "[1 photo, 2 document attached]");
-
-        let single_no_name = vec![p54_attachment(AttachmentKind::Photo)];
-        assert_eq!(
-            attachment_body_placeholder(&single_no_name),
-            "[1 photo attached]"
-        );
-    }
-
-    // ── #911 dedup-gate anchor (C0 RED) ─────────────────────────────────
-    //
-    // Locks the (A)+(B) hybrid gate contract that C1 lands at
-    // `compose_aware_inject`. Pre-C1 the predicate fn doesn't exist:
-    // these tests compile-fail at C0 = §3.10 RED anchor. C1 introduces
-    // `should_suppress_911_reinject_with_ledger` + gate-wires and the
-    // tests flip GREEN.
-    //
-    // Test names track the synthesis spec verbatim (lead-locked):
-    //   1. ledger-hit suppression
-    //   2. event-header pass-through
-    //   3. JSONL fallback on ledger MISS (closes H6 TTL eviction)
-    //   4. integration: enqueue → notify → drain → direct-inject → no PTY
-    //   5. predicate tightness: non-[AGEND-MSG] content with stray id=
-
-    #[test]
-    fn compose_aware_inject_suppressed_after_drain_for_same_msg_id_911() {
-        let home = tmp_home("911-ledger-hit");
-        let agent = "lead-911-ledger-hit";
-        let msg_id = "m-911-ledger-hit-UNIQ";
-
-        let ledger = crate::daemon::notification_dedup::Ledger::default();
-        ledger.record_inject(agent, msg_id);
-        ledger.mark_consumed(agent, msg_id);
-
-        // Canonical [AGEND-MSG] header includes the HEADER_PREFIX
-        // ANSI wrapping — match production's `format_header` output.
-        let header = format!("{HEADER_PREFIX} from=test id={msg_id} kind=task size=42");
-        let suppressed = should_suppress_911_reinject_with_ledger(&home, agent, &header, &ledger);
-
-        assert!(
-            suppressed,
-            "consumed ledger entry MUST suppress reinject (fast path B)"
-        );
-
-        std::fs::remove_dir_all(&home).ok();
-    }
-
-    #[test]
-    fn compose_aware_inject_event_header_not_gated_by_dedup_911() {
-        let home = tmp_home("911-event-pass");
-        let agent = "lead-911-event-pass";
-
-        let ledger = crate::daemon::notification_dedup::Ledger::default();
-        let event_header = format_event_header("interrupt", &[("reason", "operator stop")]);
-
-        let suppressed =
-            should_suppress_911_reinject_with_ledger(&home, agent, &event_header, &ledger);
-
-        assert!(
-            !suppressed,
-            "event headers (no id=) MUST NEVER be suppressed — not message-bound"
-        );
-
-        std::fs::remove_dir_all(&home).ok();
-    }
-
-    #[test]
-    fn compose_aware_inject_suppressed_via_jsonl_fallback_when_ledger_evicted_911() {
-        let home = tmp_home("911-jsonl-fallback");
-        let agent = "lead-911-jsonl-fallback";
-        let msg_id = "m-911-jsonl-fallback-UNIQ";
-
-        // Empty ledger — simulates H6 TTL eviction (entry was recorded
-        // + consumed 10+ min ago, sweep removed it).
-        let ledger = crate::daemon::notification_dedup::Ledger::default();
-
-        // Manually seed the inbox JSONL with a drained msg.
-        let mut msg = make_msg("dev-2", "fallback body");
-        msg.id = Some(msg_id.to_string());
-        msg.read_at = Some("2026-05-18T00:00:00Z".to_string());
-        enqueue(&home, agent, msg).unwrap();
-
-        let header = format!("{HEADER_PREFIX} from=dev-2 id={msg_id} kind=task size=10");
-        let suppressed = should_suppress_911_reinject_with_ledger(&home, agent, &header, &ledger);
-
-        assert!(
-            suppressed,
-            "JSONL fallback MUST catch drained msg when ledger has no entry (closes H6 TTL race)"
-        );
-
-        std::fs::remove_dir_all(&home).ok();
-    }
-
-    #[test]
-    fn enqueue_notify_drain_then_direct_inject_no_pty_911() {
-        // Integration: real flow through enqueue + drain. The global
-        // ledger gets the `mark_consumed` callback as a side effect of
-        // `drain`. Unique agent + msg_id to avoid pollution from
-        // other tests running on the same process-singleton ledger.
-        let home = tmp_home("911-integration");
-        let agent = "lead-911-integration-UNIQ";
-        let msg_id = "m-911-integration-UNIQ";
-
-        let mut msg = make_msg("dev-2", "integration body");
-        msg.id = Some(msg_id.to_string());
-
-        enqueue(&home, agent, msg).unwrap();
-        crate::daemon::notification_dedup::global().record_inject(agent, msg_id);
-
-        let drained = drain(&home, agent);
-        assert_eq!(drained.len(), 1, "exactly one msg should drain");
-
-        let header = format_header(&drained[0]);
-
-        let suppressed = should_suppress_911_reinject_with_ledger(
-            &home,
-            agent,
-            &header,
-            crate::daemon::notification_dedup::global(),
-        );
-
-        assert!(
-            suppressed,
-            "after enqueue → record_inject → drain (which marks consumed in global ledger), \
-             a direct re-inject of the same msg's header MUST be suppressed"
-        );
-
-        std::fs::remove_dir_all(&home).ok();
-    }
-
-    #[test]
-    fn compose_aware_inject_non_agend_msg_header_pass_through_911() {
-        let home = tmp_home("911-predicate-tight");
-        let agent = "lead-911-predicate-tight";
-
-        let ledger = crate::daemon::notification_dedup::Ledger::default();
-
-        // Non-canonical prefix but contains literal `id=` substring.
-        // Could be free-form chat text quoting a msg_id. MUST NOT be gated.
-        let non_agend = "Some chat message that happens to contain id=m-fake-12345 inline.";
-
-        let suppressed = should_suppress_911_reinject_with_ledger(&home, agent, non_agend, &ledger);
-
-        assert!(
-            !suppressed,
-            "non-[AGEND-MSG] content MUST pass through gate even with stray id= substring"
-        );
-
-        std::fs::remove_dir_all(&home).ok();
-    }
-
-    // -----------------------------------------------------------------------
-    // #982 — enqueue_with_idle_hint regression suite
-    // T0  anti-regression: raw enqueue still emits NO PTY hint.
-    // T1  idle recipient receives a hint mentioning id/kind/from/inbox count.
-    // T2  composing recipient defers hint into notification_queue (NOT injected).
-    // T3  unread count in hint matches post-enqueue inbox state.
-    // T4  same msg.id supplied by caller is preserved (no double-assignment).
-    // T5  enqueue failure path skips the hint emit (best-effort contract).
-    // T6  hint prefix uses the dedicated PENDING_HEADER_PREFIX, not HEADER_PREFIX.
-    // T7  hint sanitizes control characters in from / kind fields.
-    // T8  unique msg.id assigned when caller leaves it None.
-    // T9  emitted hint contains canonical `(use inbox tool)` affordance.
-    // T10 enqueue with `from: "from:<agent>"` strips the redundant "from:" prefix.
-    // T11 successive enqueues each carry their own distinct msg.id in the hint.
-    // T12 (recipient_state × kind) matrix — idle vs composing dispatched per kind.
-    // T13 notification_queue flush regression-pin: composing→idle releases hint.
-    // T14 #986 [pr-ready-for-merge] load-bearing: helper preserves payload + hints.
-    // -----------------------------------------------------------------------
-
-    fn capture_hint() -> std::sync::Arc<Mutex<Option<String>>> {
-        std::sync::Arc::new(Mutex::new(None))
-    }
-
-    #[test]
-    fn t0_raw_enqueue_emits_no_pty_hint() {
-        let home = tmp_home("982-t0");
-        let captured = capture_hint();
-        // Raw enqueue path — no idle hint logic.
-        enqueue(&home, "agent1", make_msg("system:test", "raw")).expect("enqueue");
-        // capture is empty because we never invoked the helper.
-        assert!(
-            captured.lock().is_none(),
-            "raw enqueue must not emit PTY hint"
-        );
-        fs::remove_dir_all(&home).ok();
-    }
-
-    #[test]
-    fn t1_idle_recipient_receives_hint() {
-        let home = tmp_home("982-t1");
-        let captured = capture_hint();
-        let mut msg = make_msg("system:waiting_on_stale", "stale alert");
-        msg.kind = Some("waiting_on_stale".to_string());
-        let captured_clone = captured.clone();
-        enqueue_with_idle_hint_with_emitter(&home, "agent1", msg, move |hint| {
-            *captured_clone.lock() = Some(hint.to_string());
-        })
-        .expect("enqueue_with_idle_hint");
-
-        let got = captured.lock().clone().expect("hint must be emitted");
-        assert!(
-            got.starts_with(PENDING_HEADER_PREFIX),
-            "hint must use pending prefix: {got:?}"
-        );
-        assert!(
-            got.contains("kind=waiting_on_stale"),
-            "hint must carry kind: {got:?}"
-        );
-        assert!(
-            got.contains("from=system:waiting_on_stale"),
-            "hint must carry from: {got:?}"
-        );
-        assert!(
-            got.contains("inbox=1"),
-            "hint must carry inbox count: {got:?}"
-        );
-        assert!(
-            got.contains("(use inbox tool)"),
-            "hint must carry inbox affordance: {got:?}"
-        );
-        fs::remove_dir_all(&home).ok();
-    }
-
-    #[test]
-    fn t2_composing_recipient_defers_hint_via_notification_queue() {
-        // Lock the global notification_dedup so the composing→deferred
-        // hint isn't suppressed by a sibling test's ledger entry.
-        let _guard = READONLY_TEST_LOCK.lock();
-        let home = tmp_home("982-t2");
-        mark_composing(&home, "agent1");
-        let msg = make_msg("system:t2", "composing test");
-        // Use the REAL emitter path (compose_aware_inject) so we can
-        // observe notification_queue side effect.
-        enqueue_with_idle_hint_with_emitter(&home, "agent1", msg, |hint| {
-            // route_notification deferral lives inside compose_aware_inject;
-            // we invoke it directly so the gate fires under our control.
-            compose_aware_inject(&home, "agent1", hint);
-        })
-        .expect("enqueue_with_idle_hint");
-
-        assert_eq!(
-            crate::notification_queue::pending_count(&home, "agent1"),
-            1,
-            "composing recipient must defer hint into notification_queue"
-        );
-        fs::remove_dir_all(&home).ok();
-    }
-
-    #[test]
-    fn t3_unread_count_in_hint_matches_post_enqueue_state() {
-        let home = tmp_home("982-t3");
-        // Pre-seed 2 unread entries so the helper's hint should show inbox=3.
-        enqueue(&home, "agent1", make_msg("seed1", "a")).expect("seed1");
-        enqueue(&home, "agent1", make_msg("seed2", "b")).expect("seed2");
-
-        let captured = capture_hint();
-        let captured_clone = captured.clone();
-        enqueue_with_idle_hint_with_emitter(
-            &home,
-            "agent1",
-            make_msg("system:t3", "new"),
-            move |hint| {
-                *captured_clone.lock() = Some(hint.to_string());
-            },
-        )
-        .expect("enqueue_with_idle_hint");
-
-        let got = captured.lock().clone().expect("hint emitted");
-        assert!(
-            got.contains("inbox=3"),
-            "hint must reflect 3 unread: {got:?}"
-        );
-        fs::remove_dir_all(&home).ok();
-    }
-
-    #[test]
-    fn t4_caller_supplied_id_preserved() {
-        let home = tmp_home("982-t4");
-        let captured = capture_hint();
-        let mut msg = make_msg("system:t4", "preset id");
-        msg.id = Some("m-caller-supplied-123".to_string());
-
-        let captured_clone = captured.clone();
-        enqueue_with_idle_hint_with_emitter(&home, "agent1", msg, move |hint| {
-            *captured_clone.lock() = Some(hint.to_string());
-        })
-        .expect("enqueue_with_idle_hint");
-
-        let got = captured.lock().clone().expect("hint emitted");
-        assert!(
-            got.contains("id=m-caller-supplied-123"),
-            "caller-supplied id must survive: {got:?}"
-        );
-        fs::remove_dir_all(&home).ok();
-    }
-
-    #[test]
-    fn t5_enqueue_failure_skips_hint_emit() {
-        let _guard = READONLY_TEST_LOCK.lock();
-        let home = tmp_home("982-t5");
-        // Force readonly so enqueue fails.
-        DISK_READONLY.store(true, Ordering::Relaxed);
-        let emitted = std::sync::Arc::new(std::sync::atomic::AtomicBool::new(false));
-        let emitted_clone = emitted.clone();
-        let result = enqueue_with_idle_hint_with_emitter(
-            &home,
-            "agent1",
-            make_msg("system:t5", "fail"),
-            move |_hint| {
-                emitted_clone.store(true, Ordering::Relaxed);
-            },
-        );
-        DISK_READONLY.store(false, Ordering::Relaxed);
-        assert!(result.is_err(), "enqueue must propagate readonly error");
-        assert!(
-            !emitted.load(Ordering::Relaxed),
-            "hint emit must be skipped on enqueue failure"
-        );
-        fs::remove_dir_all(&home).ok();
-    }
-
-    #[test]
-    fn t6_hint_uses_pending_prefix_not_header_prefix() {
-        let home = tmp_home("982-t6");
-        let captured = capture_hint();
-        let captured_clone = captured.clone();
-        enqueue_with_idle_hint_with_emitter(
-            &home,
-            "agent1",
-            make_msg("system:t6", "prefix test"),
-            move |hint| {
-                *captured_clone.lock() = Some(hint.to_string());
-            },
-        )
-        .expect("enqueue_with_idle_hint");
-
-        let got = captured.lock().clone().expect("hint emitted");
-        assert!(
-            got.contains("[AGEND-MSG-PENDING]"),
-            "hint must use AGEND-MSG-PENDING tag: {got:?}"
-        );
-        assert!(
-            !got.contains("[AGEND-MSG] "),
-            "hint must NOT collide with canonical AGEND-MSG header: {got:?}"
-        );
-        fs::remove_dir_all(&home).ok();
-    }
-
-    #[test]
-    fn t7_hint_sanitizes_control_chars() {
-        let home = tmp_home("982-t7");
-        let captured = capture_hint();
-        let mut msg = make_msg("system:t7\nattack", "ctrl chars");
-        msg.kind = Some("kind\twith\ttabs".to_string());
-        let captured_clone = captured.clone();
-        enqueue_with_idle_hint_with_emitter(&home, "agent1", msg, move |hint| {
-            *captured_clone.lock() = Some(hint.to_string());
-        })
-        .expect("enqueue_with_idle_hint");
-
-        let got = captured.lock().clone().expect("hint emitted");
-        // Control chars become spaces — split on the post-prefix region only,
-        // because PENDING_HEADER_PREFIX contains ANSI ESC bytes by design.
-        let body = got
-            .split("[AGEND-MSG-PENDING]\x1b[0m")
-            .nth(1)
-            .expect("body present");
-        assert!(
-            !body.contains('\n') && !body.contains('\t'),
-            "control chars must be sanitized to space: body={body:?}"
-        );
-        fs::remove_dir_all(&home).ok();
-    }
-
-    #[test]
-    fn t8_msg_id_auto_assigned_when_caller_omits() {
-        let home = tmp_home("982-t8");
-        let captured = capture_hint();
-        let captured_clone = captured.clone();
-        enqueue_with_idle_hint_with_emitter(
-            &home,
-            "agent1",
-            make_msg("system:t8", "no id"),
-            move |hint| {
-                *captured_clone.lock() = Some(hint.to_string());
-            },
-        )
-        .expect("enqueue_with_idle_hint");
-
-        let got = captured.lock().clone().expect("hint emitted");
-        assert!(got.contains(" id=m-"), "auto-id starts with m-: {got:?}");
-        // Auto-id format: m-{YYYYMMDDHHMMSS.6f}-{seq}
-        assert!(
-            got.contains(" id=m-2"), // 2-prefixed year (works through 2099)
-            "auto-id must contain year prefix: {got:?}"
-        );
-        fs::remove_dir_all(&home).ok();
-    }
-
-    #[test]
-    fn t9_hint_carries_inbox_affordance() {
-        let home = tmp_home("982-t9");
-        let captured = capture_hint();
-        let captured_clone = captured.clone();
-        enqueue_with_idle_hint_with_emitter(
-            &home,
-            "agent1",
-            make_msg("system:t9", "affordance"),
-            move |hint| {
-                *captured_clone.lock() = Some(hint.to_string());
-            },
-        )
-        .expect("enqueue_with_idle_hint");
-
-        let got = captured.lock().clone().expect("hint emitted");
-        assert!(
-            got.ends_with("(use inbox tool)"),
-            "hint must end with operator-trained affordance: {got:?}"
-        );
-        fs::remove_dir_all(&home).ok();
-    }
-
-    #[test]
-    fn t10_from_prefix_stripped_in_hint() {
-        let home = tmp_home("982-t10");
-        let captured = capture_hint();
-        let captured_clone = captured.clone();
-        enqueue_with_idle_hint_with_emitter(
-            &home,
-            "agent1",
-            make_msg("from:peer-agent", "strip me"),
-            move |hint| {
-                *captured_clone.lock() = Some(hint.to_string());
-            },
-        )
-        .expect("enqueue_with_idle_hint");
-
-        let got = captured.lock().clone().expect("hint emitted");
-        assert!(
-            got.contains("from=peer-agent"),
-            "from: prefix must be stripped: {got:?}"
-        );
-        assert!(
-            !got.contains("from=from:"),
-            "must not have nested from:from:: {got:?}"
-        );
-        fs::remove_dir_all(&home).ok();
-    }
-
-    #[test]
-    fn t11_successive_enqueues_carry_distinct_ids() {
-        let home = tmp_home("982-t11");
-        let hints = std::sync::Arc::new(Mutex::new(Vec::<String>::new()));
-        for i in 0..3 {
-            let hints_clone = hints.clone();
-            enqueue_with_idle_hint_with_emitter(
-                &home,
-                "agent1",
-                make_msg("system:t11", &format!("msg {i}")),
-                move |hint| {
-                    hints_clone.lock().push(hint.to_string());
-                },
-            )
-            .expect("enqueue_with_idle_hint");
-        }
-
-        let captured = hints.lock();
-        assert_eq!(captured.len(), 3, "three hints emitted");
-        // Extract id= token from each
-        let ids: Vec<&str> = captured
-            .iter()
-            .map(|h| {
-                h.split(" id=")
-                    .nth(1)
-                    .and_then(|s| s.split(' ').next())
-                    .unwrap_or("")
-            })
-            .collect();
-        assert_eq!(
-            std::collections::HashSet::<&&str>::from_iter(ids.iter()).len(),
-            3,
-            "all three ids must be distinct: {ids:?}"
-        );
-        fs::remove_dir_all(&home).ok();
-    }
-
-    #[test]
-    fn t12_pty_state_kind_matrix() {
-        // Lead Q7: explicit (pty_state × msg_kind) matrix coverage.
-        let _guard = READONLY_TEST_LOCK.lock();
-        let home = tmp_home("982-t12");
-        let kinds = ["query", "task", "report", "update", "waiting_on_stale"];
-
-        for kind in kinds {
-            // Idle path — hint reaches the closure.
-            let captured = capture_hint();
-            let captured_clone = captured.clone();
-            let mut msg = make_msg("system:t12", "idle");
-            msg.kind = Some(kind.to_string());
-            enqueue_with_idle_hint_with_emitter(&home, "agent1", msg, move |hint| {
-                *captured_clone.lock() = Some(hint.to_string());
-            })
-            .expect("idle enqueue");
-            let got = captured.lock().clone();
-            assert!(got.is_some(), "idle path must emit for kind={kind}");
-            assert!(
-                got.unwrap().contains(&format!("kind={kind}")),
-                "hint carries kind={kind}"
-            );
-
-            // Composing path — hint deferred into notification_queue.
-            let composing_home = tmp_home(&format!("982-t12-{kind}"));
-            mark_composing(&composing_home, "agent1");
-            let mut msg = make_msg("system:t12", "composing");
-            msg.kind = Some(kind.to_string());
-            let before = crate::notification_queue::pending_count(&composing_home, "agent1");
-            enqueue_with_idle_hint_with_emitter(&composing_home, "agent1", msg, |hint| {
-                compose_aware_inject(&composing_home, "agent1", hint);
-            })
-            .expect("composing enqueue");
-            let after = crate::notification_queue::pending_count(&composing_home, "agent1");
-            assert_eq!(
-                after,
-                before + 1,
-                "composing must defer 1 hint for kind={kind}"
-            );
-            fs::remove_dir_all(&composing_home).ok();
-        }
-        fs::remove_dir_all(&home).ok();
-    }
-
-    #[test]
-    fn t13_notification_queue_flush_releases_deferred_hint() {
-        let _guard = READONLY_TEST_LOCK.lock();
-        let home = tmp_home("982-t13");
-        mark_composing(&home, "agent1");
-
-        // Defer 2 hints while composing.
-        for i in 0..2 {
-            let mut msg = make_msg("system:t13", &format!("deferred {i}"));
-            msg.kind = Some("update".to_string());
-            enqueue_with_idle_hint_with_emitter(&home, "agent1", msg, |hint| {
-                compose_aware_inject(&home, "agent1", hint);
-            })
-            .expect("deferred enqueue");
-        }
-        assert_eq!(
-            crate::notification_queue::pending_count(&home, "agent1"),
-            2,
-            "both hints deferred"
-        );
-
-        // Drain returns the deferred hints.
-        let drained = crate::notification_queue::drain(&home, "agent1");
-        assert_eq!(drained.len(), 2, "drain returns deferred hints");
-        for d in &drained {
-            assert!(
-                d.text.contains("[AGEND-MSG-PENDING]"),
-                "deferred entry is the pending hint: {:?}",
-                d.text
-            );
-        }
-        fs::remove_dir_all(&home).ok();
-    }
-
-    #[test]
-    fn t15_composing_flush_uses_submit_aware_inject() {
-        // Reviewer #999 BLOCKING gap (HEAD 1ba8e69c verdict): when the
-        // recipient was composing at enqueue time, the hint deferred
-        // into notification_queue. The pre-fix flush path
-        // (`app::flush_idle_notifications`) called
-        // `inject_notification(raw=true)` which omits the backend
-        // submit_key, leaving the hint in the prompt buffer without
-        // submitting — codex one-shots silently dropped the wake.
-        //
-        // This pin verifies the contract end-to-end through the queue:
-        //
-        // 1. composing recipient → `enqueue_with_idle_hint` defers the
-        //    PTY hint into `notification_queue`
-        // 2. drain returns the deferred entry — its text body is the
-        //    `[AGEND-MSG-PENDING]` header line that the flush path will
-        //    feed to `inject_notification_with_submit`
-        // 3. `inject_notification_with_submit` delegates to the same
-        //    INJECT-no-raw payload shape as `inject_with_submit` (no
-        //    `raw: true` → `handle_inject` defaults to false →
-        //    `inject_to_agent` appends submit_key)
-        //
-        // Structural pin: verify the post-fix wiring builds the
-        // submit-aware payload shape exactly. Identical contract test
-        // pattern to `inject_with_submit_sends_raw_false`.
-        let _guard = READONLY_TEST_LOCK.lock();
-        let home = tmp_home("982-t15");
-        mark_composing(&home, "agent1");
-        let mut msg = make_msg("system:t15", "deferred update");
-        msg.kind = Some("update".to_string());
-        enqueue_with_idle_hint_with_emitter(&home, "agent1", msg, |hint| {
-            compose_aware_inject(&home, "agent1", hint);
-        })
-        .expect("enqueue");
-
-        let drained = crate::notification_queue::drain(&home, "agent1");
-        assert_eq!(drained.len(), 1, "composing window defers the hint");
-        assert!(
-            drained[0].text.contains("[AGEND-MSG-PENDING]"),
-            "deferred entry is the pending hint: {:?}",
-            drained[0].text
-        );
-
-        // Mirror the JSON shape assertion of inject_with_submit_sends_raw_false:
-        // inject_notification_with_submit MUST NOT set raw=true. (Calling the
-        // real fn needs a daemon; we verify the contract by re-constructing
-        // the payload it would build.)
-        let with_submit_payload = serde_json::json!({
-            "method": crate::api::method::INJECT,
-            "params": {"name": "agent1", "data": &drained[0].text}
-        });
-        assert!(
-            with_submit_payload["params"]["raw"].is_null(),
-            "inject_notification_with_submit must NOT set raw=true \
-             (defaults to false → inject_to_agent → submit_key appended): \
-             {with_submit_payload}"
-        );
-        fs::remove_dir_all(&home).ok();
-    }
-
-    #[test]
-    fn t14_pr_ready_for_merge_load_bearing_emits_hint() {
-        // #986 load-bearing regression-pin: the [pr-ready-for-merge] event
-        // path must surface a PTY hint after the helper migration.
-        // Pin the wire format: hint carries kind=pr-ready-for-merge and
-        // points the recipient to inbox via the affordance affordance.
-        let home = tmp_home("982-t14-ready");
-        let captured = capture_hint();
-        let mut msg = make_msg("system:pr_state", "[pr-ready-for-merge] foo/bar@feat#990");
-        msg.kind = Some("pr-ready-for-merge".to_string());
-        msg.correlation_id = Some("foo/bar@feat#990".to_string());
-
-        let captured_clone = captured.clone();
-        enqueue_with_idle_hint_with_emitter(&home, "fixup-lead", msg, move |hint| {
-            *captured_clone.lock() = Some(hint.to_string());
-        })
-        .expect("enqueue ready-event");
-
-        let got = captured.lock().clone().expect("ready-event hint emitted");
-        assert!(
-            got.contains("kind=pr-ready-for-merge"),
-            "ready-event must carry kind tag: {got:?}"
-        );
-        assert!(
-            got.contains("(use inbox tool)"),
-            "ready-event must carry affordance: {got:?}"
-        );
-        // Underlying inbox entry stays intact (durable source of truth).
-        let drained = drain(&home, "fixup-lead");
-        assert_eq!(drained.len(), 1, "inbox entry persisted");
-        assert_eq!(drained[0].kind.as_deref(), Some("pr-ready-for-merge"));
-        fs::remove_dir_all(&home).ok();
-    }
-}
