+++ src/inbox/notify.rs
@@ -1,0 +1,413 @@
+use std::path::Path;
+
+use super::message::{BroadcastContext, InboxMessage, NotifySource};
+use super::storage;
+
+/// Size threshold for header-only PTY injection. Messages with body > this
+/// value inject only a structured header line; the full body stays in inbox.
+#[allow(dead_code)]
+pub const HEADER_SIZE_THRESHOLD: usize = 300;
+
+/// Returns true when the `AGEND_POINTER_ONLY_INJECT` feature flag is set to "1".
+/// When enabled, PTY injection uses header-only format for all messages,
+/// forcing agents to call `inbox` to read content (solves dispatch non-FIFO).
+pub fn pointer_only_inject() -> bool {
+    // G3 H1: read DaemonConfig instead of env var (thread-safe)
+    crate::daemon_config::get().pointer_only_inject
+}
+
+/// ANSI-colored header prefix for visual distinction in terminal.
+pub const HEADER_PREFIX: &str = "\x1b[44;97m[AGEND-MSG]\x1b[0m";
+
+/// #982: ANSI-colored prefix for idle-hint headers emitted alongside
+/// daemon-side inbox enqueues. Distinct from [`HEADER_PREFIX`] so
+/// recipients can distinguish the canonical message body inject from
+/// the lightweight `(use inbox tool)` pointer signalling a new
+/// pending entry. Background is yellow-on-black for contrast.
+pub const PENDING_HEADER_PREFIX: &str = "\x1b[43;30m[AGEND-MSG-PENDING]\x1b[0m";
+
+/// Plain-text system message prefix (without ANSI colors).
+/// Used for detection/matching in agent PTY output parsing.
+pub const SYSTEM_MSG_PREFIX: &str = "[AGEND-MSG]";
+
+/// Agent-to-agent message prefix.
+pub const AGENT_MSG_PREFIX: &str = "[from:";
+
+/// Sanitize a value for header inclusion: replace control chars with space.
+fn sanitize_header_value(s: &str) -> String {
+    s.chars()
+        .map(|c| if c.is_control() { ' ' } else { c })
+        .collect()
+}
+
+/// Format a single-line structured header for PTY injection.
+/// Fields: from / id / kind / thread / parent / size.
+/// Optional fields (thread/parent) omitted when None.
+#[allow(dead_code)]
+pub fn format_header(msg: &InboxMessage) -> String {
+    // #761: strip the redundant `from:` prefix that `Source::Agent`'s
+    // Display impl adds. `strip_prefix` returns `Some` for agent sources
+    // and `None` otherwise, so `system:` / `user:` namespaces survive
+    // untouched via the `unwrap_or` fallback.
+    let from_value = msg.from.strip_prefix("from:").unwrap_or(&msg.from);
+    let mut parts = vec![
+        HEADER_PREFIX.to_string(),
+        format!("from={}", sanitize_header_value(from_value)),
+    ];
+    if let Some(ref id) = msg.id {
+        parts.push(format!("id={}", sanitize_header_value(id)));
+    }
+    if let Some(ref kind) = msg.kind {
+        parts.push(format!("kind={}", sanitize_header_value(kind)));
+    }
+    if let Some(ref thread) = msg.thread_id {
+        parts.push(format!("thread={}", sanitize_header_value(thread)));
+    }
+    if let Some(ref parent) = msg.parent_id {
+        parts.push(format!("parent={}", sanitize_header_value(parent)));
+    }
+    parts.push(format!("size={}", msg.text.chars().count()));
+    if !msg.attachments.is_empty() {
+        let paths: Vec<&str> = msg
+            .attachments
+            .iter()
+            .map(|a| a.path.to_str().unwrap_or("?"))
+            .collect();
+        parts.push(format!("attachments=[{}]", paths.join(",")));
+    }
+    // Sprint 54 layer-5 broadcast visibility: surface broadcast routing in
+    // the PTY-inject header.
+    if let Some(ref ctx) = msg.broadcast_context {
+        parts.push(format!("broadcast={}", ctx.count));
+        if let Some(ref team) = ctx.team {
+            parts.push(format!("team={}", sanitize_header_value(team)));
+        }
+    }
+    if let Some(ref excerpt) = msg.in_reply_to_excerpt {
+        parts.push(format!(
+            "reply_to_excerpt={}",
+            sanitize_header_value(excerpt)
+        ));
+    }
+    parts.join(" ")
+}
+
+/// Build the `in_reply_to_excerpt` value for inbound replies.
+///
+/// Returns `None` when `text` is empty so callers can `and_then` over an
+/// optional reply target. Otherwise truncates to 200 chars (Unicode-safe via
+/// `chars().take(...)`) and appends `…` if the original exceeded 200 chars.
+pub(crate) fn build_excerpt(text: &str, author: &str) -> Option<String> {
+    if text.is_empty() {
+        return None;
+    }
+    let truncated: String = text.chars().take(200).collect();
+    let ellipsis = if text.chars().count() > 200 {
+        "…"
+    } else {
+        ""
+    };
+    Some(format!("[{author}] {truncated}{ellipsis}"))
+}
+
+/// Format a single-line event header (non-message events like poll-reminder).
+/// Uses the same ANSI prefix as [`format_header`] for visual consistency.
+pub fn format_event_header(kind: &str, fields: &[(&str, &str)]) -> String {
+    let mut parts = vec![
+        HEADER_PREFIX.to_string(),
+        format!("kind={}", sanitize_header_value(kind)),
+    ];
+    for (k, v) in fields {
+        parts.push(format!("{}={}", k, sanitize_header_value(v)));
+    }
+    parts.join(" ")
+}
+
+/// Deliver a message: always enqueue to inbox JSONL for persistence,
+/// then inject to PTY (inline or pointer-only depending on feature flag).
+pub fn deliver(
+    home: &Path,
+    agent_name: &str,
+    source: &NotifySource<'_>,
+    text: &str,
+    _submit_key: &str,
+    kind: Option<String>,
+    broadcast_context: Option<BroadcastContext>,
+) {
+    let msg = InboxMessage {
+        from: source.to_string(),
+        text: text.to_string(),
+        kind,
+        timestamp: chrono::Utc::now().to_rfc3339(),
+        broadcast_context,
+        ..Default::default()
+    };
+    let _ = storage::enqueue(home, agent_name, msg);
+    notify_agent(home, agent_name, source, text);
+}
+
+/// Sprint 54 silent-drop layer-4 hotfix: aggregate kind counts for the
+/// PTY-inject header `attachments=[…]` field.
+pub(crate) fn summarize_attachments_for_header(
+    attachments: &[crate::channel::event::Attachment],
+) -> Option<String> {
+    if attachments.is_empty() {
+        return None;
+    }
+    use crate::channel::event::AttachmentKind;
+    let mut counts: [usize; 5] = [0; 5];
+    for a in attachments {
+        let i = match a.kind {
+            AttachmentKind::Photo => 0,
+            AttachmentKind::Voice => 1,
+            AttachmentKind::Document => 2,
+            AttachmentKind::Video => 3,
+            AttachmentKind::Sticker => 4,
+        };
+        counts[i] += 1;
+    }
+    let labels = ["photo", "voice", "document", "video", "sticker"];
+    let parts: Vec<String> = counts
+        .iter()
+        .zip(labels.iter())
+        .filter(|(n, _)| **n > 0)
+        .map(|(n, label)| format!("{n} {label}"))
+        .collect();
+    Some(parts.join(", "))
+}
+
+/// Sprint 54 silent-drop layer-4 hotfix: human-readable body
+/// placeholder used when the inbox message has no text but attachments
+/// are present.
+pub(crate) fn attachment_body_placeholder(
+    attachments: &[crate::channel::event::Attachment],
+) -> String {
+    if attachments.is_empty() {
+        return String::new();
+    }
+    let summary = summarize_attachments_for_header(attachments).unwrap_or_default();
+    if attachments.len() == 1 {
+        if let Some(name) = &attachments[0].original_filename {
+            return format!("[{summary}: {name}]");
+        }
+    }
+    format!("[{summary} attached]")
+}
+
+/// Pure function: build the notification string that will be injected to PTY.
+/// `pointer_only=true` → header-only (no body); `false` → inline text.
+pub fn format_notification_for_inject(
+    pointer_only: bool,
+    source: &NotifySource<'_>,
+    text: &str,
+    attachments: &[crate::channel::event::Attachment],
+) -> String {
+    let attach_summary = summarize_attachments_for_header(attachments);
+    if pointer_only {
+        let mut header = format!(
+            "[{source}] [AGEND-MSG] size={} (use inbox tool)",
+            text.len()
+        );
+        if let Some(s) = &attach_summary {
+            header.push_str(&format!(" attachments=[{s}]"));
+        }
+        header.push_str(&source.reply_hint());
+        header
+    } else {
+        // Body-replace path: empty text + non-empty attachments would
+        // produce a content-less inline notification pre-r0. Substitute
+        // the placeholder so the agent at minimum sees what kind of
+        // media is waiting.
+        let display_text = if text.is_empty() && !attachments.is_empty() {
+            attachment_body_placeholder(attachments)
+        } else if text.chars().count() > 200 {
+            let truncated: String = text.chars().take(200).collect();
+            format!("{truncated}... (use the inbox MCP tool to read full message)")
+        } else {
+            text.to_string()
+        };
+        format!("[{source}] {display_text}{}", source.reply_hint())
+    }
+}
+
+pub fn notify_agent(home: &Path, agent_name: &str, source: &NotifySource<'_>, text: &str) {
+    notify_agent_with_attachments(home, agent_name, source, text, &[]);
+}
+
+/// Sprint 54 silent-drop layer-4 hotfix: variant of `notify_agent` that
+/// carries attachment metadata into the PTY-inject formatter.
+pub fn notify_agent_with_attachments(
+    home: &Path,
+    agent_name: &str,
+    source: &NotifySource<'_>,
+    text: &str,
+    attachments: &[crate::channel::event::Attachment],
+) {
+    // Sprint 24 P1 (F-NEW-DAEMON-HEALTH-CLASSIFIER-1): record this
+    // central inject point so the daemon health classifier can
+    // distinguish "idle waiting (no input pending)" from "hung
+    // unresponsive (input pending past last response)".
+    crate::daemon::heartbeat_pair::update_with(agent_name, |p| {
+        p.last_input_at_ms = crate::daemon::heartbeat_pair::now_ms();
+    });
+    let notification =
+        format_notification_for_inject(pointer_only_inject(), source, text, attachments);
+    compose_aware_inject(home, agent_name, &notification);
+    // #836: record the (agent, msg_id) tuple in the dedup ledger
+    if let Some(msg_id) =
+        crate::daemon::notification_dedup::extract_msg_id_from_header(&notification)
+    {
+        crate::daemon::notification_dedup::global().record_inject(agent_name, &msg_id);
+    }
+    // Store raw body AFTER inject
+    crate::daemon::heartbeat_pair::update_with(agent_name, |p| {
+        p.last_input_text = Some(text.to_string());
+    });
+}
+
+/// Compose-aware notification delivery: checks `is_composing` and enqueues
+/// if the target agent is mid-typing, otherwise injects **with** submit_key
+/// so idle agents actually wake up on the incoming notification.
+pub fn compose_aware_inject(home: &Path, agent_name: &str, notification: &str) {
+    // #911 dedup gate
+    if should_suppress_911_reinject_with_ledger(
+        home,
+        agent_name,
+        notification,
+        crate::daemon::notification_dedup::global(),
+    ) {
+        return;
+    }
+    let _ = route_notification(home, agent_name, notification, |msg| {
+        inject_with_submit(home, agent_name, msg)
+    });
+}
+
+/// #911 dedup gate predicate. Hybrid (A)+(B) suppression check for
+/// a candidate PTY notification.
+pub(crate) fn should_suppress_911_reinject_with_ledger(
+    home: &Path,
+    agent_name: &str,
+    notification: &str,
+    ledger: &crate::daemon::notification_dedup::Ledger,
+) -> bool {
+    // Reviewer condition C: only canonical `HEADER_PREFIX` headers gate
+    if !notification.starts_with(HEADER_PREFIX) {
+        return false;
+    }
+    let Some(msg_id) = crate::daemon::notification_dedup::extract_msg_id_from_header(notification)
+    else {
+        return false;
+    };
+    // Fast path (B): in-memory ledger.
+    if ledger.should_suppress_reinject(agent_name, &msg_id) {
+        tracing::info!(
+            agent = %agent_name,
+            msg_id = %msg_id,
+            "#911 compose_aware_inject suppressed: dedup ledger hit"
+        );
+        return true;
+    }
+    // Fallback (A): JSONL read_at source-of-truth.
+    if storage::msg_already_drained_in_jsonl(home, agent_name, &msg_id) {
+        tracing::info!(
+            agent = %agent_name,
+            msg_id = %msg_id,
+            "#911 compose_aware_inject suppressed: JSONL fallback (ledger MISS, read_at set)"
+        );
+        return true;
+    }
+    false
+}
+
+/// #982 Direction A: persist a message to the recipient's inbox AND emit a
+/// best-effort `[AGEND-MSG-PENDING]` PTY hint so daemon-side events do not
+/// strand silently when the recipient is at an idle prompt.
+pub fn enqueue_with_idle_hint(home: &Path, target: &str, msg: InboxMessage) -> anyhow::Result<()> {
+    enqueue_with_idle_hint_with_emitter(home, target, msg, |hint| {
+        compose_aware_inject(home, target, hint);
+    })
+}
+
+/// Test-seam variant of [`enqueue_with_idle_hint`]. Accepts a closure
+/// that receives the formatted hint string so unit tests can verify
+/// the wire format without standing up the API loopback.
+pub(crate) fn enqueue_with_idle_hint_with_emitter<F>(
+    home: &Path,
+    target: &str,
+    mut msg: InboxMessage,
+    emit_hint: F,
+) -> anyhow::Result<()>
+where
+    F: FnOnce(&str),
+{
+    // Pre-stamp the id so the hint can reference it. enqueue() leaves an
+    // existing id untouched, so this stays a single auto-assignment.
+    storage::ensure_msg_id(&mut msg);
+    let id = msg.id.clone().unwrap_or_default();
+    let from = msg.from.clone();
+    let kind = msg.kind.clone();
+
+    storage::enqueue(home, target, msg)?;
+
+    // Build the PTY hint AFTER the inbox write succeeds
+    let pending = storage::unread_count(home, target).0;
+    let from_short = from.strip_prefix("from:").unwrap_or(&from);
+    let kind_str = kind.as_deref().unwrap_or("");
+    let hint = format!(
+        "{} id={} kind={} from={} inbox={} (use inbox tool)",
+        PENDING_HEADER_PREFIX,
+        sanitize_header_value(&id),
+        sanitize_header_value(kind_str),
+        sanitize_header_value(from_short),
+        pending,
+    );
+    emit_hint(&hint);
+    Ok(())
+}
+
+fn inject_with_submit(home: &Path, agent_name: &str, message: &str) -> anyhow::Result<()> {
+    let resp = crate::api::call(
+        home,
+        &serde_json::json!({
+            "method": crate::api::method::INJECT,
+            "params": {"name": agent_name, "data": message}
+        }),
+    )?;
+    if resp["ok"].as_bool() == Some(true) {
+        Ok(())
+    } else {
+        anyhow::bail!(
+            "{}",
+            resp["error"]
+                .as_str()
+                .unwrap_or("inject with submit failed")
+        );
+    }
+}
+
+/// #982 RC: submit-aware notification inject for the
+/// `notification_queue` flush path.
+pub fn inject_notification_with_submit(
+    home: &Path,
+    agent_name: &str,
+    notification: &str,
+) -> anyhow::Result<()> {
+    inject_with_submit(home, agent_name, notification)
+}
+
+pub(super) fn route_notification<F>(
+    home: &Path,
+    agent_name: &str,
+    notification: &str,
+    mut injector: F,
+) -> anyhow::Result<()>
+where
+    F: FnMut(&str) -> anyhow::Result<()>,
+{
+    if crate::notification_queue::is_composing(home, agent_name) {
+        crate::notification_queue::enqueue(home, agent_name, notification)?;
+        return Ok(());
+    }
+    injector(notification)
+}
